Default outputs of run_fredi()

A dataframe containing the default outputs of run_fredi()

defaultResults

Format

A data frame with 173 016 rows and 18 columns:

sector

Name of the sector

variant

Name of the adaptation or variant (values are sector-specific)

impactYear

Name of the impact year ("2010", "2090", "N/A", or "Interpolation")

impactType

Name of the impact type ("all" or sector-specific values)

region

Name of the associated region ("Midwest", "Northeast", "Northern Plains", "Northwest", "Southeast", "Southern Plains", "Southwest", or "National Total")

model_type

Type of model used to calculate impacts --- either "GCM" (for "Global Climate Model") or "SLR" for ("Sea Level Rise")

model

Name of the GCM ("Average", "CanESM2", "CCSM4", "GCM Ensemble", "GFDL-CM3", "GISS-E2-R", "HadGEM2-ES", "MIROC5", "MRI-CGCM3") or SLR model ("Interpolation")

year

Year of the output

driverValue

Value for the associated driver type (in "degrees Celsius" for temperature and in "cm" for sea level rise)

driverUnit

Unit for the associated driver type ("degrees Celsius" for temperature and "cm" for sea level rise)

driverType

Associated driver type ("Temperature" or "GMSL (SLR)")

includeaggregate

A 0 or 1 value indicating whether the variant is included in aggregation

sectorprimary

A 0 or 1 value indicating whether the sector is a primary CIRA sector

gdp_usd

National gross domestic product (GDP) for associated year in 2015 USD (U.S. dollars)

national_pop

National population for associated year

gdp_per_cap

National GDP per capita for associated year in 2015 USD (U.S. dollars) per capita

reg_pop

Regional population for associated year

annual_impacts

Annual impacts for associated year, region, sector, variant, impact year, impact type, and model

...

...

Source

https://epa.gov/cira/FrEDI/