Six driver scenarios that can be passed as inputs to run_fredi() and run_fredi_sv()

gcamScenarios is a data frame object containing six driver scenarios that can be passed as inputs to run_fredi() and run_fredi_sv(). This data frame has five columns – year, temp_C_global, temp_C_conus, slr_cm, scenario, and model:

Usage

gcamScenarios

Format

A data frame with 606 rows and 6 columns:

year

Year

temp_C_global

Global temperatures, in degrees Celsius, for the associated year and scenario

temp_C_conus

Temperatures for the contiguous U.S. (CONUS), in degrees Celsius, for the associated year and scenario (converted from global temperatures using convertTemps(from="global"))

slr_cm

Global Mean Sea Level Rise (GMSL or SLR), in centimeters, for the associated year and scenario (calculated from global temperatures using temps2slr())

scenario

Associated scenario identifier (e.g., "ECS_3.0_REF")

model

A string ("Hector_GCAM_v5.3") identifying the model (Hector, with GCAM v5.3) used in generating the global temperatures associated with each scenario

Source

https://epa.gov/cira/FrEDI/

Details

year

The Common Era (CE) year associated with the observation.

temp_C_global

Global temperature (i.e., degrees of warming above the baseline year of 1995), in degrees Celsius, for the associated year and scenario.

temp_C_conus

Temperatures (i.e., degrees of warming above the baseline year of 1995) for the contiguous U.S. (CONUS), in degrees Celsius, for the associated year and scenario (calculated from global temperatures using convertTemps(from="global")).

slr_cm

Global Mean Sea Level Rise (GMSL or SLR), in centimeters, for the associated year and scenario (calculated from global temperatures using temps2slr()).

scenario

Associated scenario identifier (e.g., "ECS_3.0_REF").

model

A string ("Hector_GCAM_v5.3") identifying the model (Hector, with GCAM v5.3) used in generating the global temperatures associated with each scenario.

The scenarios in this data frame were created using Hector with GCAM v5.3:

Hector is an open-source, reduced-form model (Hartin et al., 2015) used to model temperatures associated with emissions scenarios .

GCAM v5.3 – an open source model that represents the linkages between physical and economic systems (Calvin et al., 2019).

Scenario identifiers in the scenario column of gcamScenarios() have the string "ECS_3.0_REF_" as a prefix, followed by a suffix indicating an emissions intensity associated with the scenario (e.g., "20") – for instance, the default scenario for run_fredi() and run_fredi_sv() is "ECS_3.0_REF". Other scenarios include "ECS_3.0_REF_20", "ECS_3.0_REF_30", "ECS_3.0_REF_50", "ECS_3.0_REF_70", and "ECS_3.0_REF_90".

Users can use the scenarios in gcamScenarios as inputs to FrEDI or the FrEDI SV Module:

Users can filter the gcamScenarios data frame to any of these six scenarios, which can then be passed directly to the run_fredi() function via a named element (temp and/or slr) in a list passed to the inputsList argument – e.g., run_fredi(inputsList=list(temp=gcamScenarios |> filter(scenario=="ECS_3.0_REF"))) will run using the default temperature scenario.

Any or all of the GCAM scenarios can be passed directly to the FrEDI SV module via the run_fredi_sv() function via a named list element (temp and/or slr) in a list passed to the inputsList argument – e.g., run_fredi_sv(inputsList=list(temp=gcamScenarios)) will run the SV module for all the GCAM scenarios provided in gcamScenarios.

gcamScenarios can also be combined with other provided scenarios (gdpDefault, popDefault) in function calls (e.g., run_fredi(inputsList=list(temp=gcamScenarios |> filter(scenario=="ECS_3.0_REF"), gdp=gdpDefault, pop=popDefault))) or user-provided data frames. For more information, see documentation for run_fredi(), FrEDI:run_fredi_sv(), FrEDI:run_fredi_ghg(), and import_inputs().

Calvin, K., Patel, P., Clarke, L., et al. 2019. https://doi.org/10.5194/gmd-12-677-2019.

Hartin, C.A., Patel, P., Schwarber, A., Link, R.P. and Bond-Lamberty, B.P., 2015. URL: <www.geosci-model-dev-discuss.net/7/7075/2014/>. doi:10.5194/gmdd-7-7075-2014.