Create plots of FrEDI outputs

This function creates plots for the summarized FrEDI outputs. get_plots() returns a list with heatmaps for model types present in the data (GCMs and SLR scenarios) and ribbon plots, by sector.

get_plots(
  data = NULL,
  column = "annual_impacts",
  undiscounted = TRUE,
  plotTypes = "all",
  save = FALSE,
  directory = NULL,
  groupVars = c("sector", "variant")
)

Arguments

data

Data frame of summarized outputs produced by run_fredi(). Do not change column names of the run_fredi() output before running get_plots().

column

A character string indicating the name of the column in the data for which to create plots.

undiscounted

A TRUE/FALSE value indicating whether the values in column represent undiscounted values or discounted values (i.e., present values). Defaults to undiscounted=TRUE.

plotTypes

Character string or character vector indicating which types of plots to produce. Options are c("heatmaps", "ribbon", "all"). Set plotTypes="all" (default) to produce both types of plots.

save

A TRUE/FALSE value indicating whether to save results. If a directory value is supplied (i.e., !is.null(directory)), defaults to save=TRUE. Otherwise, default is save=FALSE.

directory

A character string indicating the location of a directory in which to save the plot objects. No default (i.e., directory=NULL).

groupVars

A character vector indicating columns to use for grouping. Defaults to groupVars=c("sector", "variant")

Value

get_plots() returns a nested list of plots:

List Index	Description

heatmaps	List of heatmaps with list elements for all unique values for model types present in data (i.e., "GCM" for results calculated with temperature as the driver value and "SLR" for results calculated with GMSL as the driver value.

ribbon	List of lists of annual results. List of annual plots contains a list of sectors. Each sector contains a list of nested regional plots (regional, with all regions except national results and national, for national results).

Details

This function creates plots for the summarized FrEDI outputs. get_plots() returns a list with heatmaps for model types present in the data (GCMs and SLR scenarios) and annual results for all sectors and variants. Results from FrEDI must be summed across impact types before using get_plots() (use run_fredi() with the defaults, use run_fredi(aggLevels= "impacttype"), or run aggregate_impacts(aggLevels="impactType") on the output from run_fredi()).

By default, get_plots() plots results from the "annual_impacts" column (or users can specify a column name in the data with column).

The argument undiscounted is used by get_plots() for plot labels and in file and directory names for saving results.

get_plots() produces heatmaps (plotTypes="heatmaps") for the outputs of FrEDI and/or plots the average value and range of impacts as a time series (plotTypes="ribbon") for each sector-variant-region combination. Users can specify which plot types to produce by setting plotTypes. Set plotTypes="all" (default) to produce both heat maps and annual results or specify a single type (plotTypes="heatmaps" and plotTypes="ribbon", respectively).

The heatmaps display the numeric values in the specified column (e.g., "annual_impacts") as a grid of colored pixels. Each row in the grid corresponds to a sector-variant combination (e.g., "Coastal Properties, No Adaptation"), while columns in the grid correspond to years. In other words, the heatmaps display the relative intensity of the impacts of a sector and variant compared to others. The colors in the heatmap are a gradient ranging from dark blue (impacts with values below zero) to dark red (impacts with values above zero), with a midpoint at zero (missing values appear as grey pixels). The scale of the gradient is determined from the underlying data, with the darkest points corresponding to the minimum and maximum values.

If temperature-driven (GCM sectors) and SLR-driven (SLR sectors) sectors are both present in the data, get_plots() will produce a separate heatmap for each. Each heatmap displays panels for each region (stacked vertically) and underlying models (organized horizontally).

Setting (plotTypes="ribbon") plots the annual impacts as time series. For temperature-driven sectors, the model average is plotted as a line and the range of model values (minimum and maximum) are plotted as a ribbon plot. For the SLR-driven sectors, the interpolated impacts are plotted as a line. For sectors with multiple variants, impacts for individual variants are displayed in separate panels (organized horizontally).

If save=TRUE and the user supplies a path to a directory (i.e., !is.null(directory)), get_plots() will try to save results to the specified directory. Separate directories are created within the specified directory for heatmaps and ribbon plots.

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Create temperature binning scenario
df_tempExOut <- get_fredi(aggLevels="none", pv=TRUE, silent=TRUE)
#> Error in get_fredi(aggLevels = "none", pv = TRUE, silent = TRUE): could not find function "get_fredi"

### Aggregate temperature binning summary across multiple columns
agg_tempExOut <- df_tempExOut %>% aggregate_impacts(columns=c("annual_impacts", "discounted_impacts"))
#> Error in ungroup(.): object 'df_tempExOut' not found

### Create list of plots for aggregated results
agg_plotList <- agg_tempExOut %>% get_plots()
#> Error in (function (y) {    y$valueColumn <- y[, column]    return(y)})(.): object 'agg_tempExOut' not found

### Create list of heatmaps for regional values only:
reg_plotList <- agg_tempExOut %>% filter(region!="National Total") %>% get_plots(plotTypes="heatmaps")
#> Error in filter(., region != "National Total"): object 'agg_tempExOut' not found

### Create list of annual plots for national values only:
nation_plotList <- agg_tempExOut %>% filter(region=="National Total") %>% get_plots(plotTypes="annual")
#> Error in filter(., region == "National Total"): object 'agg_tempExOut' not found