This function provides information about the sectors in FrEDI.

This helper function returns a character vector of sector names (default) or a data frame of sectors with related information (e.g., associated variants, impact types, etc.), which the user can supply to the run_fredi() sectorList argument.

get_sectorInfo(description = F, gcmOnly = F, slrOnly = F)

Arguments

description

= FALSE. Logical value indicating whether to include information about each sector. Returns a dataframe if description=TRUE and returns a character vector of sector names if description=FALSE (default).

gcmOnly

= FALSE. Logical value indicating whether to return only sectors with climate impacts modeled using global climate model (GCM) results.

slrOnly

= FALSE. Logical value indicating whether to return only sectors with climate impacts modeled using sea level rise (SLR) scenarios.

Value

If description=FALSE (default), outputs a character vector containing the names of sectors available for FrEDI.

If description=TRUE, `, outputs a dataframe containing the names of sectors available for FrEDI in one column, with information about the sector model type, variants, impact years, and impact types in the remaining columns.

Details

If description=FALSE (default), this helper function returns a character vector of sector names, which the user can supply to the sectorList argument to run_fredi(). If description=TRUE, get_sectorInfo() returns dataframe of sectors with related information returns a dataframe containing the sectors available for FrEDI along with additional information. Sector names are in the first column, with additional columns for the associated model type ("GCM" or "SLR"), variants, impact years, and impact types in the remaining columns. Variants, impact years, and impact types vary by sector.

Users can specify whether to return only GCM sectors or SLR sectors by setting gcmOnly=TRUE or slrOnly=TRUE, respectively. get_sectorInfo() will return the sectors in the form specified by description (see above).

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Return a character vector with the names of all of the sectors in FrEDI:
get_sectorInfo()
#>  [1] "ATS Extreme Temperature"                  
#>  [2] "Air Quality"                              
#>  [3] "Asphalt Roads"                            
#>  [4] "CIL Agriculture"                          
#>  [5] "CIL Crime"                                
#>  [6] "CIL Extreme Temperature"                  
#>  [7] "Coastal Properties"                       
#>  [8] "Electricity Demand and Supply"            
#>  [9] "Electricity Transmission and Distribution"
#> [10] "Extreme Temperature"                      
#> [11] "High Tide Flooding and Traffic"           
#> [12] "Inland Flooding"                          
#> [13] "Labor"                                    
#> [14] "Marine Fisheries"                         
#> [15] "Rail"                                     
#> [16] "Roads"                                    
#> [17] "Southwest Dust"                           
#> [18] "Urban Drainage"                           
#> [19] "Valley Fever"                             
#> [20] "Water Quality"                            
#> [21] "Wildfire"                                 
#> [22] "Wind Damage"                              
#> [23] "Winter Recreation"                        

### Return a dataframe of all of the sectors in FrEDI (sector names and additional information)
get_sectorInfo(description=T, gcmOnly=T)
#>                                       sector model_type
#> 3                    ATS Extreme Temperature        GCM
#> 1                                Air Quality        GCM
#> 2                              Asphalt Roads        GCM
#> 4                            CIL Agriculture        GCM
#> 5                                  CIL Crime        GCM
#> 6                    CIL Extreme Temperature        GCM
#> 8              Electricity Demand and Supply        GCM
#> 9  Electricity Transmission and Distribution        GCM
#> 10                       Extreme Temperature        GCM
#> 12                           Inland Flooding        GCM
#> 13                                     Labor        GCM
#> 14                          Marine Fisheries        GCM
#> 15                                      Rail        GCM
#> 16                                     Roads        GCM
#> 17                            Southwest Dust        GCM
#> 18                            Urban Drainage        GCM
#> 19                              Valley Fever        GCM
#> 20                             Water Quality        GCM
#> 21                                  Wildfire        GCM
#> 22                               Wind Damage        GCM
#> 23                         Winter Recreation        GCM
#>                                                     variants impactYears
#> 3    Mean, Low Confidence Interval, High Confidence Interval         N/A
#> 1                             2011 Emissions, 2040 Emissions         N/A
#> 2                                                        N/A         N/A
#> 4          With CO2 Fertilization, Without CO2 Fertilization         N/A
#> 5                                                        N/A         N/A
#> 6  Median, Low Confidence Interval, High Confidence Interval         N/A
#> 8                                                        N/A         N/A
#> 9   No Adaptation, Proactive Adaptation, Reactive Adaptation         N/A
#> 10                                 Adaptation, No Adaptation  2010, 2090
#> 12                                                       N/A         N/A
#> 13                                                       N/A         N/A
#> 14                                                       N/A         N/A
#> 15  No Adaptation, Proactive Adaptation, Reactive Adaptation         N/A
#> 16  No Adaptation, Proactive Adaptation, Reactive Adaptation         N/A
#> 17                                                       N/A  2010, 2090
#> 18                                                       N/A         N/A
#> 19                                                       N/A         N/A
#> 20                                                       N/A         N/A
#> 21                                                       N/A         N/A
#> 22                                                       N/A         N/A
#> 23                                                       N/A  2010, 2090
#>                                                                                   impactTypes
#> 3                                                                                         N/A
#> 1                                                                                Ozone, PM2.5
#> 2                                                                                         N/A
#> 4                                                               Cotton, Maize, Soybean, Wheat
#> 5                                                                           Property, Violent
#> 6                                                                                         N/A
#> 8                                                                                         N/A
#> 9                                                                                         N/A
#> 10                                                                                  Cold, Hot
#> 12                                                                                        N/A
#> 13                                                                                        N/A
#> 14                                                                                        N/A
#> 15                                                                                        N/A
#> 16                                                                                        N/A
#> 17 Acute Myocardial Infarction, All Cardiovascular, All Mortality, All Respiratory, Asthma ER
#> 18                                                                                        N/A
#> 19                                                           Lost Wages, Morbidity, Mortality
#> 20                                                                                        N/A
#> 21                                                       Morbidity, Mortality, Response Costs
#> 22                                                                                        N/A
#> 23                                          Alpine Skiing, Cross-Country Skiing, Snowmobiling

### Return a character vector with only the names of the temperature-driven sectors:
get_sectorInfo(gcmOnly=T)
#>  [1] "ATS Extreme Temperature"                  
#>  [2] "Air Quality"                              
#>  [3] "Asphalt Roads"                            
#>  [4] "CIL Agriculture"                          
#>  [5] "CIL Crime"                                
#>  [6] "CIL Extreme Temperature"                  
#>  [7] "Electricity Demand and Supply"            
#>  [8] "Electricity Transmission and Distribution"
#>  [9] "Extreme Temperature"                      
#> [10] "Inland Flooding"                          
#> [11] "Labor"                                    
#> [12] "Marine Fisheries"                         
#> [13] "Rail"                                     
#> [14] "Roads"                                    
#> [15] "Southwest Dust"                           
#> [16] "Urban Drainage"                           
#> [17] "Valley Fever"                             
#> [18] "Water Quality"                            
#> [19] "Wildfire"                                 
#> [20] "Wind Damage"                              
#> [21] "Winter Recreation"                        

### Run FrEDI for only the temperature-driven sectors and view results:
df_x <- run_fredi(sectorList=get_sectorInfo(gcmOnly=T))
#> Checking input values...
#> Creating temperature scenario from defaults...
#> Creating SLR scenario from temperature scenario...
#> Creating GDP scenario from defaults...
#> Creating Population scenario from defaults...
#> Formatting results...
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...
#> 
#> Finished.