Retrieve a character vector of sectors available in the FrEDI Social Vulnerability (SV) module (run_fredi_sv) or a data frame with SV sectors and additional information.

get_sv_sectorInfo returns a character vector with the names of sectors in the FrEDI Social Vulnerability (SV) module (default) or a data frame of SV sectors with additional information (e.g., associated variants, model type, etc.).

Usage

get_sv_sectorInfo(description = F, gcmOnly = F, slrOnly = F)

Arguments

description

Logical value indicating whether to include information about each sector. Returns a data frame if description=TRUE and returns a character vector of sector names if description=FALSE (default).

gcmOnly

Logical value indicating whether to return only sectors with climate impacts modeled using global climate model (GCM) results.

slrOnly

Logical value indicating whether to return only sectors with climate impacts modeled using sea level rise (SLR) scenarios.

Value

get_sv_sectorInfo() will return SV sectors in the form specified by description:

If description=FALSE (default), outputs a character vector containing the names of sectors available for the FrEDI SV Module.

If description=TRUE, outputs a data frame containing the names of sectors available for the FrEDI SV Module in one column, with information about the sector model type, model type ("GCM" or "SLR") and associated driver unit ("degrees Celsius" or "cm", respectively ), impact units (e.g., mortality, etc.), and available variants in the remaining columns.

Details

get_sv_sectorInfo returns a character vector with the names of sectors in the FrEDI SV module (description=FALSE, default) or a data frame with SV sectors and additional information (description=TRUE).

If description=FALSE (default), get_sv_sectorInfo returns a character vector of names of sectors that can be passed to the run_sv_fredi() sector argument. Specify whether to return only GCM sectors by running get_sv_sectorInfo(gcmOnly=TRUE) or SLR sectors by running get_sv_sectorInfo(slrOnly=TRUE).

If description=TRUE, get_sv_sectorInfo returns a data frame of sectors with related information, such as whether a particular sector is driven primarily by temperature (modelType="GCM") or sea level rise (modelType="SLR"), associated driver units ("degrees Celsius" for temperature-driven sectors, "cm" for SLR-driven sectors), impact units (e.g., mortality, etc.), and sector variants. Users can use get_sv_sectorInfo to determine which sectors can be passed to the run_sv_fredi() sector argument and/or to determine whether a particular sector is driven primarily by temperature (modelType="GCM") or sea level rise (modelType="SLR").

Users can specify whether to return only GCM sectors by setting gcmOnly=TRUE or SLR sectors by setting slrOnly=TRUE.

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Return a character vector with the names of all of the sectors in the FrEDI SV Module:
get_sv_sectorInfo()
#> [1] "Air Quality - Childhood Asthma"                
#> [2] "Air Quality - Premature Mortality"             
#> [3] "Coastal Properties"                            
#> [4] "Extreme Temperature"                           
#> [5] "Labor"                                         
#> [6] "Roads"                                         
#> [7] "Transportation Impacts from High Tide Flooding"

### Return a data frame of all of the sectors in the FrEDI SV Module (sector names and additional information)
get_sv_sectorInfo(description=T)
#>                                           sector modelType      driverUnit
#> 1                 Air Quality - Childhood Asthma       GCM degrees Celsius
#> 2              Air Quality - Premature Mortality       GCM degrees Celsius
#> 3                             Coastal Properties       SLR              cm
#> 4                            Extreme Temperature       GCM degrees Celsius
#> 5                                          Labor       GCM degrees Celsius
#> 6                                          Roads       GCM degrees Celsius
#> 7 Transportation Impacts from High Tide Flooding       SLR              cm
#>                                        impactUnit
#> 1                          Childhood asthma cases
#> 2                             Premature Mortality
#> 3 Individuals threatened with total property loss
#> 4                                       Mortality
#> 5                             Hours of labor lost
#> 6                                  Hours of delay
#> 7                                  Hours of delay
#>                              variants
#> 1                                 N/A
#> 2                                 N/A
#> 3 Without Adaptation, With Adaptation
#> 4                                 N/A
#> 5                                 N/A
#> 6 No Adaptation, Proactive Adaptation
#> 7 Without Adaptation, With Adaptation

### Return a character vector with only the names of the temperature-driven sectors:
get_sv_sectorInfo(gcmOnly=T)
#> [1] "Air Quality - Childhood Asthma"    "Air Quality - Premature Mortality"
#> [3] "Extreme Temperature"               "Labor"                            
#> [5] "Roads"                            

### Return a character vector with only the names of the temperature-driven sectors:
get_sv_sectorInfo(slrOnly=T)
#> [1] "Coastal Properties"                            
#> [2] "Transportation Impacts from High Tide Flooding"