Project annual average climate change impacts throughout the 21st century for available sectors

This function allows users to project annual average climate change impacts through 2300 (2010-2300) for available sectors (see get_sectorInfo()). Users may specify an optional list of custom scenarios. The output is an R data frame object containing annual average impacts, by year, for each sector, variant, impact type, model (GCM or SLR scenario), and region.

Usage

run_fredi(
  inputsList = list(tempInput = NULL, slrInput = NULL, gdpInput = NULL, popInput = NULL),
  sectorList = NULL,
  aggLevels = c("national", "modelaverage", "impactyear", "impacttype"),
  elasticity = 0.4,
  maxYear = 2090,
  thru2300 = FALSE,
  pv = FALSE,
  baseYear = 2010,
  rate = 0.03,
  outputList = FALSE,
  silent = TRUE
)

Arguments

inputsList

A list of named elements named elements (names(inputsList)= c("tempInput", "slrInput", "gdpInput", "popInput")), each containing data frames of custom temperature, global mean sea level rise (GMSL), gross domestic product (GDP), and/or population scenarios, respectively, over a continuous period in the range 2010 to 2300. Note: temperature and sea level rise inputs should start in 2000 or earlier. Temperature and sea level rise inputs should start in 2000 or earlier. Values for population and GDP scenarios can start in 2010 or earlier. Values for each scenario type must be within reasonable ranges. For more information, see import_inputs().

sectorList

A character vector indicating a selection of sectors for which to calculate results (see get_sectorInfo()). If NULL, all sectors are included.

aggLevels

Levels of aggregation at which to summarize data: one or more of c("national", "modelaverage", "impactyear", "impacttype", "all", "none"). Defaults to all levels (i.e., aggLevels="all"). Uses the same aggregation levels as aggregate_impacts().

thru2300

A TRUE/FALSE shortcut that overrides the maxYear argument to run the model to 2300.

pv

A TRUE/FALSE value indicating whether to calculate present values for the annual impacts. Defaults to pv=FALSE. Present values (i.e., discounted impacts) are calculated as discounted_impacts=annual_impacts/(1+rate)^(year-baseYear). Set an annual discounting rate and a base year using baseYear and rate.

baseYear

Base year used for calculating present values of annual impacts (i.e., discounting). Defaults to baseYear=2010.

rate

Annual discount rate used in calculating present values of annual impacts (i.e., discounting). Defaults to rate=0.03 (i.e., 3% per year).

outputList

A TRUE/FALSE value indicating whether to output results as a data frame object (outputList=FALSE, default) or to return a list of objects (outputList=TRUE) that includes information about model provenance (including input arguments and input scenarios) along with the data frame of results.

silent

A TRUE/FALSE value indicating the level of messaging desired by the user (default=TRUE).

elasticity=0.4

A numeric value indicating an elasticity to use for adjusting VSL for applicable sectors and impacts (defaults to elasticity=0.4). Applicable sectors and impacts are Air Quality (all impact types), ATS Extreme Temperature, CIL Extreme Temperature, Extreme Temperature (all impact types), Mental Health, Southwest Dust (All Mortality), Valley Fever (Mortality), Vibrio, and Wildfire (Mortality).

maxYear=2090

A numeric value indicating the maximum year for the analysis.

Value

If outputList=FALSE, the output of run_fredi() is a data frame object containing annual average impacts over the analysis period, for each sector, variant, model (GCM or SLR scenario), impact type, impact year, and region. If outputList=TRUE, run_fredi() returns a list object containing the following:

arguments, containing a list with values for the arguments passed to run_fredi() (with the exception of scenarios passed to the inputsList argument, which are provided in the driverScenarios list element)

driverScenarios, a list object containing elements with the driver scenarios for temperature, SLR, population, and GDP used in the model

results, containing a data frame of annual average impacts

Details

This function allows users to project annual average climate change impacts through 2300 (2010-2300) for available sectors. run_fredi() is the main function in the FrEDI R package, described elsewhere (See https://epa.gov/cira/FrEDI for more information).

Users can specify an optional list of custom scenarios with inputsList (for more information on the format of inputs, see import_inputs()). run_fredi() uses default scenarios for temperature, population, and GDP when no inputs are specified (i.e., inputsList=NULL) or for empty elements of the inputs list. If the user does not specify an input scenario for GMSL (i.e., inputsList=list(slrInput=NULL), run_fredi() first converts the CONUS temperature scenario to global temperatures and then converts the global temperatures to a global mean sea level rise (GMSL) height in centimeters. For more information on the conversion of CONUS temperatures to global temperatures, see convertTemps(). For more information on the conversion of global temperatures to GMSL, see temps2slr().

Temperature and GMSL inputs must begin in 2000 or earlier, whereas values for population and GDP scenarios can start in 2010 or earlier. Values for input scenarios must be within reasonable ranges (for instance, negative values for population and GDP are non-sensical). If a user inputs a custom scenario with values outside the allowable ranges, run_fredi() will not run the scenarios and will instead stop and return an error message. For more information, see import_inputs().

The input temperature scenario (passed to run_fredi() via the inputsList argument) requires temperatures for CONUS in degrees Celsius relative to 1995 (degrees of warming relative to the baseline). Temperature values must be greater than or equal to zero degrees Celsius (CONUS temperatures). Users can convert global temperatures to CONUS temperatures using FrEDI::convertTemps(from="global") or by specifying FrEDI::import_inputs(temptype="global") when importing a temperature scenario from a CSV file.

Values for the sea level rise (SLR) scenario are for global mean sea level rise (GMSL) must be in centimeters (cm) and values must be greater than or equal to zero and less than or equal to 250 cm.

Population and gross domestic product (GDP) values must be greater than or equal to zero.

If inputsList=NULL, run_fredi() uses defaults for all scenarios. Otherwise, run_fredi() looks for a list object passed to the argument inputsList. Within that list, run_fredi() looks for list elements tempInput, slrInput, gdpInput, and popInput containing data frames with custom scenarios for temperature, GMSL, GDP, and regional population, respectively. run_fredi() will default back to the default scenarios for any list elements that are NULL or missing. In other words, running run_fredi(inputsList=list()) returns the same outputs as running run_fredi(). For help importing custom scenarios from CSV files, refer to the pre-processing function import_inputs().

run_fredi() linearly interpolates missing annual values for all input scenarios using non-missing values (requires at least two non-missing values). Temperatures are interpolated using 1995 as the baseline year (i.e., the central year of the 1986-2005 baseline). In other words, the temperature (in degrees Celsius) is set to zero for the year 1995 and GMSL is set to zero for the year 2000. The interpolated temperature and GMSL scenarios are combined into a column called driverValue, along with additional columns for year, the driver unit (column "driverUnit", with driverUnit="degrees Celsius" and driverUnit="cm" for temperature- and SLR-driven sectors, respectively), and the associated model type (column "model_type", with model_type="GCM" and model_type="SLR" for temperature- and SLR-driven sectors, respectively).

The population scenario must provide annual regional values for population, with national totals calculated from regional values. FrEDI uses the national population scenario and the GDP scenario to calculate GDP per capita. Values for regional population, national population, national GDP (in 2015$), and national per capita GDP (in 2015$/capita) are provided in the results data frame in columns "reg_pop", "national_pop", "gdp_usd", and "gdp_percap", respectively.

By default, run_fredi() will calculate impacts for all sectors included in the tool. Alternatively, users can pass a character vector specifying a single sector or a subset of sectors using the sectorList argument. To see a list of sectors included within FrEDI, run get_sectorInfo(). If sectorList=NULL (default), all sectors are included.

run_fredi() aggregates or summarizes results to level(s) of aggregation specified by the user (passed to aggLevels) using the post-processing helper function aggregate_impacts(). Users can specify a single aggregation level or multiple aggregation levels by passing a single character string or character vector to aggLevels. Options for aggregation include calculating national totals (aggLevels="national"), averaging across model types and models (aggLevels="modelaverage"), summing over all impact types (aggLevels="impacttype"), and interpolating between impact year estimates (aggLevels="impactYear"). Users can specify all aggregation levels at once by specifying aggLevels="all" (default) or no aggregation levels (aggLevels="none").

For each of the aggLevels, run_fredi() performs the following calculations using aggregate_impacts() (note that the "variant" column referred to below contains information about the variant (or adaptation) name or “N/A”, as applicable):

Aggregation Level	Description

national	Annual values are summed across all regions present in the data. I.e., data is grouped by columns "sector", "variant", "impactType", "impactYear", "model_type", "model", and "year") and summed across regions. Years which have missing column data for all regions return as NA. The rows of the data frame of national values (with column region="National Total") are then added as rows to the results.

modelaverage	For temperature-driven sectors, annual results are averaged across all GCM models present in the data. I.e., data is grouped by columns "sector", "variant", "impactType", "impactYear", "model_type", "region", and "year" and averaged across models (SLR impacts are estimated as an interpolation between SLR scenarios). Averages exclude missing values. Years which have missing column data for all models return as NA. The rows of model averages (with column model="Average" are then added as rows to the results data frame.

impacttype	Annual results are summed across all impact types by sector present in the data. I.e., data is grouped by columns "sector", "variant", "impactType", "impactYear","model_type", "model", "region", and "year" and summed across impact types. Mutates column impactType="all" for all values. Years which have missing column data for all impact types return as NA. If results are aggregated across impact types, information about physical impacts (columns "physicalmeasure" and "physical_impacts") are dropped.

impactyear	Annual results for sectors with only one impact year estimate (i.e., impactYear = "N/A") are separated from those with multiple impact year estimates. Sectors with multiple impact years have separate results for impact years 2010 and 2090. For these sectors, annual results are linearly interpolated between impact year estimates. For any model run years above 2090, annual results for sectors with multiple impact years return the 2090 estimate. The interpolated values are bound back to the results for sectors with a single impact year estimate, and column impactYear set to impactYear="Interpolation" for all values.

Annual impacts for each sector, variant, impact type, and impact year combination included in the model are calculated by multiplying scaled climate impacts by a physical scalar and economic scalars and multipliers. Some sectors use Value of a Statistical Life (VSL) to adjust the value non-linearly over time. run_fredi() uses a default value of elasticity=0.4to adjust VSL for applicable sectors and impacts. Applicable sectors and impacts are Air Quality (all impact types), ATS Extreme Temperature, CIL Extreme Temperature, Extreme Temperature (all impact types), Mental Health, Southwest Dust (All Mortality), Valley Fever (Mortality), Vibrio, and Wildfire (Mortality). A custom elasticity can be passed to the elasticity argument; to keep VSL constant over time, specify elasticity=1.

By default, run_fredi() calculates impacts starting in the year 2010 and ending in 2090. Specify an alternative end year for the analysis using the maxYear argument (defaults to maxYear=2090). The minimum and maximum valid values for maxYear are maxYear=2011 and maxYear=2300, respectively. Alternatively, run the model through the year 2300 by specifying thru2300=TRUE (this will override the maxYear argument and set maxYear=2300). Note that the default scenarios included within FrEDI stop in the year 2090; to get non-zero/non-missing values for years after 2090, users must specify a maxYear after 2090 and also provide custom input scenarios out to the desired end year.

Users can choose to calculate present values of annual impacts (i.e., discounted impacts), by setting pv=TRUE (defauts to pv=FALSE). If pv=TRUE, discounted impacts are calculated using a base year and annual discount rate as discounted_impacts=annual_impacts/(1+rate)^(year-baseYear). Set base year and annual discount rate using baseYear (defaults to baseYear=2010) and rate (defaults to 3% i.e., rate=0.03), respectively.

run_fredi() defaults to returning a data frame of annual average impacts over the analysis period, for each sector, variant, model (GCM or SLR scenario), impact type, impact year, and region (outputList=FALSE). If outputList=TRUE, run_fredi() returns a list object containing information about values for function arguments and driver scenarios in addition to the data frame of impacts

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Load FrEDI
require(FrEDI)

### Run function with defaults (same as `defaultResults` dataset)
run1 <- run_fredi()
#> Checking input values...
#> No temperature scenario provided...using default temperature scenario...
#> Creating SLR scenario from temperature scenario...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from defaults...
#> Formatting results...
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...
#> 
#> Finished.

### Load climate scenarios and glimpse data
data("gcamScenarios")
gcamScenarios %>% glimpse
#> Rows: 546
#> Columns: 4
#> $ year     <dbl> 2000, 2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009, 2~
#> $ temp_C   <dbl> 0.1804171, 0.2001162, 0.2198154, 0.2395145, 0.2592137, 0.2789~
#> $ slr_cm   <dbl> 0.0000000, 0.2660846, 0.5362304, 0.8104140, 1.0886122, 1.3708~
#> $ scenario <chr> "ECS_3.0_ref_0", "ECS_3.0_ref_0", "ECS_3.0_ref_0", "ECS_3.0_r~

### Load population scenario and glimpse data
data(popScenario)
popScenario %>% glimpse
#> Rows: 133
#> Columns: 3
#> $ year    <dbl> 2000, 2005, 2010, 2015, 2020, 2025, 2030, 2035, 2040, 2045, 20~
#> $ region  <chr> "Midwest", "Midwest", "Midwest", "Midwest", "Midwest", "Midwes~
#> $ reg_pop <dbl> 58596051, 59678411, 60760771, 62531700, 64097103, 65655846, 66~

### Subset climate scenario
temps1 <- gcamScenarios %>% filter(scenario=="ECS_3.0_ref_0") %>% select(year, temp_C)

### Run custom scenario
run2 <- run_fredi(inputsList=list(tempInput=temps1, popInput=popScenario))
#> Checking input values...
#> Creating temperature scenario from user inputs...
#> Creating SLR scenario from temperature scenario...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from user inputs...
#> Formatting results...
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...
#> 
#> Finished.

### Load scenarios from file:
scenariosPath <- system.file(package="FrEDI") %>% file.path("extdata","scenarios")
scenariosPath %>% list.files
#> [1] "GCAM_scenario.csv" "pop_scenario.csv"  "slr_from_GCAM.csv"

### Temperature Scenario File Name
tempInputFile <- scenariosPath %>% file.path("GCAM_scenario.csv")

### SLR Scenario File Name
slrInputFile  <- scenariosPath %>% file.path("slr_from_GCAM.csv")

### Population Scenario File Name
popInputFile  <- scenariosPath %>% file.path("pop_scenario.csv")

### Import inputs
x_inputs <- import_inputs(
  tempfile = tempInputFile,
  slrfile  = slrInputFile,
  popfile  = popInputFile
)
#> 
#> In import_inputs():
#> 	User supplied temperature input...
#> 		Importing data from C:/Users/emcduf01/OneDrive - Environmental Protection Agency (EPA)/Documents/R/R-4.0.3/library/FrEDI/extdata/scenarios/GCAM_scenario.csv...
#> 		Formatting temperature inputs...
#> 	User supplied SLR input...
#> 		Importing data from C:/Users/emcduf01/OneDrive - Environmental Protection Agency (EPA)/Documents/R/R-4.0.3/library/FrEDI/extdata/scenarios/slr_from_GCAM.csv...
#> 		Formatting SLR inputs...
#> 	User supplied regional population input...
#> 		Importing data from C:/Users/emcduf01/OneDrive - Environmental Protection Agency (EPA)/Documents/R/R-4.0.3/library/FrEDI/extdata/scenarios/pop_scenario.csv...
#> 		Formatting regional population inputs...
#> 			User specified `popform='wide'`...Gathering population by region...
#> Finished.

### Run custom scenarios
run3 <- run_fredi(inputsList=x_inputs)
#> Checking input values...
#> Creating temperature scenario from user inputs...
#> Creating SLR scenario from user inputs...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from user inputs...
#> Formatting results...
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...
#> 
#> Finished.

### Get information on sectors:
get_sectorInfo()
#>  [1] "ATS Extreme Temperature"                  
#>  [2] "Air Quality"                              
#>  [3] "Asphalt Roads"                            
#>  [4] "CIL Agriculture"                          
#>  [5] "CIL Crime"                                
#>  [6] "CIL Extreme Temperature"                  
#>  [7] "Coastal Properties"                       
#>  [8] "Electricity Demand and Supply"            
#>  [9] "Electricity Transmission and Distribution"
#> [10] "Extreme Temperature"                      
#> [11] "High Tide Flooding and Traffic"           
#> [12] "Inland Flooding"                          
#> [13] "Labor"                                    
#> [14] "Marine Fisheries"                         
#> [15] "Mental Health"                            
#> [16] "Rail"                                     
#> [17] "Roads"                                    
#> [18] "Southwest Dust"                           
#> [19] "Urban Drainage"                           
#> [20] "Valley Fever"                             
#> [21] "Vibrio"                                   
#> [22] "Water Quality"                            
#> [23] "Wildfire"                                 
#> [24] "Wind Damage"                              
#> [25] "Winter Recreation"                        

### Run for a single sector, with default inputs, no aggregation, and elasticity=1:
run4 <- run_fredi(sectorList="ATS Extreme Temperature", aggLevels="none", elasticity=1)
#> Checking input values...
#> No temperature scenario provided...using default temperature scenario...
#> Creating SLR scenario from temperature scenario...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from defaults...
#> Formatting results...
#> 
#> Finished.

### Set end year for analysis to 2100 using default scenarios (values after 2090 will all be missing, since default scenarios only have values out to 2090)
run5 <- run_fredi(maxYear=2100)
#> Checking input values...
#> No temperature scenario provided...using default temperature scenario...
#> Creating SLR scenario from temperature scenario...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from defaults...
#> Formatting results...
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...
#> 
#> Finished.

### Set end year for analysis to 2300 using default scenarios (values after 2090 will all be missing, since default scenarios only have values out to 2090)
run6 <- run_fredi(thru2300=TRUE)
#> Checking input values...
#> No temperature scenario provided...using default temperature scenario...
#> Creating SLR scenario from temperature scenario...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from defaults...
#> Formatting results...
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...
#> 
#> Finished.

### Run FrEDI, calculate discounted impacts, and return output list
run7 <- run_fredi(pv=TRUE, baseYear=2020, rate=0.3, outputList=TRUE)
#> Checking input values...
#> No temperature scenario provided...using default temperature scenario...
#> Creating SLR scenario from temperature scenario...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from defaults...
#> Formatting results...
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...
#> 
#> Finished.