Calculate annual average impacts from temperature and sea level change with distribution among different populations throughout the 21st century for available sectors

run_fredi_sv allows users to project annual average impacts from temperature and sea level change throughout the 21st century (2010-2100) with distribution among different populations for available sectors. Users can run run_fredi_sv() for individual sectors to generate annual physical impacts for SV populations. run_fredi_sv() can be run with default population, temperature, and sea level trajectories or using custom trajectories. The output of run_fredi_sv() is an R data frame object containing annual average physical impacts at five-year increments for the period 2010 to 2100. The basic structure, specific methodology, and underlying data supporting FrEDI-SV are derived from EPA’s independently peer-reviewed September 2021 report.

Usage

run_fredi_sv(
  sector = NULL,
  inputsList = list(pop = NULL, temp = NULL, slr = NULL),
  silent = TRUE,
  .testing = FALSE
)

Arguments

sector

A character string indicating the sector for which the FrEDI SV module should calculate impacts (see get_sv_sectorInfo() for a list of available sectors).

silent

A logical (TRUE/FALSE) value indicating the level of messaging desired by the user (defaults to silent=TRUE).

inputsList=list(pop=NULL, temp=NULL, slr=NULL)

A list with named elements (pop, temp, and/or slr), each containing data frames of custom scenarios for state-level population, temperature, and/or global mean sea level rise (GMSL) trajectories, respectively, over a continuous period. Temperature and sea level rise inputs should start in 2000 or earlier. Values for population scenarios can start in 2010 or earlier. Values for each scenario type must be within reasonable ranges. For more information, see import_inputs().

pop. The input population scenario requires a data frame object with a single scenario of population values for each of the 48 U.S. states and the District of Columbia comprising the contiguous U.S. (CONUS).

The population scenario must have five columns with names "region", "state", "postal", "year", and "_pop" containing the NCA region name ("Midwest", "Northeast", "Northern Plains", "Northwest", "Southeast", "Southern Plains", or "Southwest"), the state name, the two-letter postal code abbreviation for the state (e.g., "ME" for Maine), the year, and the state population, respectively.

The input population scenario can only contain a single scenario, in contrast to values for temperature or SLR inputs. In other words, run_fredi_sv() uses the same population scenario when running any and all of the temperature or SLR scenarios passed to run_fredi_sv().

If the user does not specify an input scenario for population (i.e., popInput = NULL, run_fredi_sv() uses a default population scenario (see documentation for popScenario).

Population inputs must have at least one non-missing value in 2010 or earlier and at least one non-missing value in or after the final analysis year (2100).

Population values must be greater than or equal to zero.

temp or slr. The input temperature or SLR scenario should be a data frame containing one or more custom scenarios. These inputs should be formulated similarly to those for run_fredi() and import_inputs(), with an additional column (scenario) indicating the unique scenario identifier. Temperature and/or SLR input scenarios must have at least one non-missing value in the year 2000 or earlier and at least one non-missing value in or after the final analysis year (2100).

temp. Temperature inputs are used by run_fredi_sv() with temperature-driven sectors; run get_sv_sectorInfo(gcmOnly=TRUE) to get a list of the temperature-driven sectors available for the SV module. Temperature inputs require a data frame with columns of year, temp_C, and scenario, respectively containing the year associated with an observation, temperature values for CONUS in degrees Celsius of warming relative to a 1995 baseline (where 1995 is the central year of a 1986-2005 baseline period -- i.e., values should start at zero in the year 1995), and a unique scenario identifier. If no temperature scenario is specified (i.e., inputsList$temp is NULL) when running a temperature-driven sector, run_fredi_sv() will use a default temperature scenario (see FrEDI:gcamScenarios).

slr. SLR inputs are used by run_fredi_sv() with sea level rise-driven sectors; run get_sv_sectorInfo(slrOnly=TRUE) to get a list of the SLR-driven sectors available for the SV module. SLR inputs require a data frame with columns of year, slr_cm, and scenario, respectively containing the global mean sea level rise in centimeters relative to a 2000 baseline (i.e., values should start at zero in the year 2000), and a unique scenario identifier. If no SLR scenario is specified (i.e., inputsList$slr is NULL) when running a temperature-driven sector: if a user has supplied a temperature scenario (i.e., inputsList$temp is not NULL), run_fredi_sv() will calculate sea level rise values from the temperature inputs using the temps2slr() function; if no temperature scenario is provided, run_fredi_sv will use a default SLR scenario (see FrEDI:gcamScenarios).

Value

The output of run_fredi_sv() is an R data frame object containing average annual physical impacts, with distribution among different populations, at the NCA region level and five-year increments.

Details

run_fredi_sv() projects annual average impacts from temperature and sea level change throughout the 21st century (2010-2100) for available sectors, with distribution among different populations, using default or user-specified population, temperature, and sea level rise (SLR) trajectories. run_fredi_sv() is the main function for the FrEDI SV module in the FrEDI R package, described elsewhere (See https://epa.gov/cira/FrEDI for more information). The SV module extends the FrEDI framework to different populations using data underlying a 2021 U.S. Environmental Protection Agency (EPA) report.

Users can run run_fredi_sv() to generate annual physical impacts for SV groups for individual sectors. When running run_fredi_sv(), users must specify one of the sectors in the SV module; use get_sv_sectorInfo() for a list of available sectors.

run_fredi_sv() can be run with default population, temperature, and sea level trajectories or use run_fredi_sv() to run custom scenarios. Running run_fredi_sv() with custom temperature and sea level scenarios requires passing a data frame of scenarios to the driverInput argument. run_fredi_sv() can also be run with a custom population scenario by passing a data frame of state population trajectories to the popInput argument; unlike the temperature and sea level scenarios, run_fredi_sv() will only run a single population scenario at a time.

The output of run_fredi_sv() is an R data frame object containing average annual physical impacts, with distribution among different populations, at the NCA region level and five-year increments.

References

Bierwagen, B., D. M. Theobald, C. R. Pyke, A. Choate, P. Groth, J. V. Thomas, and P. Morefield. 2010. https://doi.org/10.1073/pnas.1002096107.

EPA. 2017. https://doi.org/10.13140/RG.2.2.14466.79045

EPA. 2021. Technical Documentation on the Framework for Evaluating Damages and Impacts (FrEDI). U.S. Environmental Protection Agency, EPA 430-R-21-004. Available at https://epa.gov/cira/FrEDI/.

EPA. 2021. https://www.epa.gov/cira/social-vulnerability-report/.

United Nations. 2015. World population prospects: The 2015 revision. New York: United Nations, Department of Economic and Social Affairs, Population Division.

U.S. Global Change Research Program. 2015. https://scenarios.globalchange.gov/regions_nca4.

Examples