Summarize and aggregate impacts from FrEDI (calculate national totals, average across models, sum impact types, and interpolate between impact year estimates)

Summarize and aggregate impacts from FrEDI (calculate national totals, average across models, sum impact types, and interpolate between impact estimate years).

Usage

aggregate_impacts(
  data,
  aggLevels = c("national", "modelaverage", "impactyear", "impacttype"),
  columns = c("annual_impacts"),
  groupByCols = c("sector", "variant", "impactYear", "impactType", "model_type", "model",
    "region"),
  mode = "all"
)

Arguments

data

Data frame of results FrEDI (outputs from run_fredi())

aggLevels

Levels of aggregation at which to summarize data: one or more of c("national", "modelAverage", "impactYear", "impactType", "all"). Defaults to all levels (i.e., aggLevels="all").

columns

Character vector of columns for which to aggregate results (defaults to columns=c("annual_impacts")).

groupByCols

Character vector indicating which columns to use for grouping. Defaults to groupByCols=c("sector", "variant", "impactYear", "impactType", "model_type", "model", "region"). Note that the "variant" column referred to below contains information about the variant (or adaptation) name or “N/A”, as applicable.

Value

The output of run_fredi() is an R data frame object containing annual average impacts, by year (2010-2090), summarized at the specified aggregation and grouping levels.

Details

This post-processing helper function aggregates and summarizes the FrEDI results to levels of aggregation specified by the user (passed to aggLevels). Users can specify a single aggregation level or multiple aggregation levels by passing a single character string or character vector to aggLevels. Options for aggregation include calculating national totals (aggLevels="national"), averaging across model types and models (aggLevels="modelAverage"), summing over all impact types (aggLevels="impactType"), and interpolating between impact year estimates (aggLevels="impactYear"). Users can specify all aggregation levels at once by specifying aggLevels="all" (default). This post-processing helper function aggregates and summarizes temperature binning results to levels of aggregation specified by the user (passed to aggLevels). Users can specify a single aggregation level or multiple aggregation levels by passing a single character string or character vector to aggLevels. Options for aggregation include calculating national totals (aggLevels="national"), averaging across model types and models (aggLevels="modelaverage"), summing over all impact types (aggLevels="impacttype"), and interpolate between impact year estimates (aggLevels="impactyear"). Users can specify all aggregation levels at once by specifying aggLevels="all" (default) or no aggregation levels (aggLevels="none").

Before aggregating impacts for national totals and/or model averages, aggregate_impacts() will drop any pre-summarized results (i.e., values for which region="National Total" and/or for which model="average", respectively) that are already present in the data `and then reaggregate at those levels.

For each of the aggLevels, aggregate_impacts() performs the following calculations (note that the "variant" column referred to below contains information about the variant (or adaptation) name or “N/A”, as applicable):

Aggregation Level	Description

national	Annual values are summed across all regions present in the data. I.e., data is grouped by columns "sector", "variant", "impactType", "impactYear", "model_type", "model", and "year") and summed across regions. Years which have missing column data for all regions return as NA. The rows of the data frame of national values (with column region="National Total") are then added as rows to the results.

modelaverage	For temperature-driven sectors, annual results are averaged across all GCM models present in the data. I.e., data is grouped by columns "sector", "variant", "impactType", "impactYear", "model_type", "region", and "year" and averaged across models (SLR impacts are estimated as an interpolation between SLR scenarios). Averages exclude missing values. Years which have missing column data for all models return as NA. The rows of model averages (with column model="Average" are then added as rows to the results data frame.

impacttype	Annual results are summed across all impact types by sector present in the data. I.e., data is grouped by columns "sector", "variant", "impactType", "impactYear","model_type", "model", "region", and "year" and summed across impact types. Mutates column impactType="all" for all values. Years which have missing column data for all impact types return as NA. If results are aggregated across impact types, information about physical impacts (columns "physicalmeasure" and "physical_impacts") are dropped.

impactyear	Annual results for sectors with only one impact year estimate (i.e., impactYear = "N/A") are separated from those with multiple impact year estimates. Sectors with multiple impact years have separate results for impact years 2010 and 2090. For these sectors, annual results are linearly interpolated between impact year estimates. For any model run years above 2090, annual results for sectors with multiple impact years return the 2090 estimate. The interpolated values are bound back to the results for sectors with a single impact year estimate, and column impactYear set to impactYear="Interpolation" for all values.

Note that aggregate_impacts() drops columns not used in grouping or aggregation.

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Create temperature binning scenario
df_tempExOut <- run_fredi(aggLevels="none", pv=TRUE, silent=TRUE)
#> Checking input values...
#> No temperature scenario provided...using default temperature scenario...
#> Creating SLR scenario from temperature scenario...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from defaults...
#> Formatting results...
#> 
#> Finished.

### Aggregate temperature binning summary across multiple columns
agg_tempExOut <- df_tempExOut %>% aggregate_impacts(columns=c("annual_impacts", "discounted_impacts"))
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...