Default outputs of run_fredi() A dataframe containing the default outputs of run_fredi()

Default outputs of run_fredi() A dataframe containing the default outputs of run_fredi()

Usage

defaultResults

Format

A data frame with 1,501,500 rows and 20 columns:

sector

Name of the associated sector.

variant

Name of the associated variant or adaptation (values are sector-specific).

impactYear

Name of the impact year ("2010", "2090", "N/A", or "Interpolation").

impactType

Name of the impact type ("all" or sector-specific values).

region

Name of the associated region ("Midwest", "Northeast", "Northern Plains", "Northwest", "Southeast", "Southern Plains", "Southwest", or "National Total").

state

Name of the associated state (or District of Columbia): "Alabama", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", "Delaware", "District of Columbia", "Florida", "Georgia", "Idaho", "Illinois", "Indiana", "Iowa", "Kansas", "Kentucky", "Louisiana", "Maine", "Maryland", "Massachusetts", "Michigan", "Minnesota", "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada", "New Hampshire", "New Jersey", "New Mexico", "New York", "North Carolina", "North Dakota", "Ohio", "Oklahoma", "Oregon", "Pennsylvania", "Rhode Island", "South Carolina", "South Dakota", "Tennessee", "Texas", "Utah", "Vermont", "Virginia", "Washington", "West Virginia", "Wisconsin", "Wyoming". For the "National Total" region, state = "All". For sectors with only has region-level data, state = "N/A".

postal

Postal code abbreviation of the associated state (e.g., postal="AL" for state="Alabama"). The postal code abbreviation for the District of Columbia is "DC". For the "National Total" region, postal = "US". For sectors with only has region-level data, postal = "N/A".

model_type

Type of model used to calculate impacts --- either "GCM" (for "Global Climate Model") or "SLR" for ("Sea Level Rise")

model

Name of the GCM ("Average", "CanESM2", "CCSM4", "GCM Ensemble", "GFDL-CM3", "GISS-E2-R", "HadGEM2-ES", "MIROC5", "MRI-CGCM3") or SLR model ("Interpolation")

sectorprimary

A 0 or 1 value indicating whether the sector is a primary CIRA sector (and whether it should be included when summing across sectors). sectorprimary=1 if a primary sector and sectorprimary=0 if not.

includeaggregate

A 0 or 1 value indicating whether the variant is the primary variant for the sector (and whether it should be included when summing across sectors). includeaggregate=1 if the variant is a primary variant and includeaggregate=0 if not).

driverType

Associated driver type ("Temperature" or "GMSL (SLR)").

driverUnit

Unit for the associated driver type ("degrees Celsius" for temperature and "cm" for sea level rise).

driverValue

Value for the associated driver type (in "degrees Celsius" for temperature and in "cm" for sea level rise).

gdp_usd

National gross domestic product (GDP) for associated year in 2015 USD (U.S. dollars).

national_pop

National population for associated year.

gdp_per_cap

National GDP per capita for associated year in 2015 USD (U.S. dollars) per capita.

state_pop

Population for the associated state (for state-level data) or region (for region-only data).

annual_impacts

Annual impacts for associated sector, variant, impact type, impact year, region, state, model type, model, and year.

Source

https://epa.gov/cira/FrEDI/