This function provides information about the sectors in FrEDI.

This helper function returns a character vector of sector names (default) or a data frame of sectors with related information (e.g., associated variants, impact types, etc.), which the user can supply to the run_fredi() sectorList argument.

Usage

get_sectorInfo(description = F, gcmOnly = F, slrOnly = F)

Arguments

description

= FALSE. Logical value indicating whether to include information about each sector. Returns a dataframe if description=TRUE and returns a character vector of sector names if description=FALSE (default).

gcmOnly

= FALSE. Logical value indicating whether to return only sectors with impacts driven by temperature change.

slrOnly

= FALSE. Logical value indicating whether to return only sectors with impacts driven by sea level change.

Value

If description=FALSE (default), outputs a character vector containing the names of sectors available for FrEDI.

If description=TRUE, `, outputs a dataframe containing the names of sectors available for FrEDI in one column, with information about the sector model type, variants, impact years, and impact types in the remaining columns.

Details

If description=FALSE (default), this helper function returns a character vector of sector names, which the user can supply to the sectorList argument to run_fredi(). If description=TRUE, get_sectorInfo() returns dataframe of sectors with related information returns a dataframe containing the sectors available for FrEDI along with additional information. Sector names are in the first column, with additional columns for the associated model type ("GCM" or "SLR"), variants, impact years, and impact types in the remaining columns. Variants, impact years, and impact types vary by sector.

Users can specify whether to return only GCM sectors or SLR sectors by setting gcmOnly=TRUE or slrOnly=TRUE, respectively. get_sectorInfo() will return the sectors in the form specified by description (see above).

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Return a character vector with the names of all of the sectors in FrEDI:
get_sectorInfo()
#>  [1] "ATS Temperature-Related Mortality"             
#>  [2] "Asphalt Roads"                                 
#>  [3] "CIL Agriculture"                               
#>  [4] "CIL Crime"                                     
#>  [5] "CIL Temperature-Related Mortality"             
#>  [6] "Climate-Driven Changes in Air Quality"         
#>  [7] "Coastal Properties"                            
#>  [8] "Electricity Demand and Supply"                 
#>  [9] "Electricity Transmission and Distribution"     
#> [10] "Extreme Temperature"                           
#> [11] "Forestry Surplus Loss"                         
#> [12] "Inland Flooding"                               
#> [13] "Labor"                                         
#> [14] "Learning Loss"                                 
#> [15] "Lyme Disease"                                  
#> [16] "Marine Fisheries"                              
#> [17] "Marsh Migration (Primary)"                     
#> [18] "Marsh Migration (Secondary)"                   
#> [19] "Outdoor Recreation"                            
#> [20] "Rail"                                          
#> [21] "Roads"                                         
#> [22] "Southwest Dust"                                
#> [23] "Suicide"                                       
#> [24] "Transportation Impacts from High Tide Flooding"
#> [25] "Urban Drainage"                                
#> [26] "Valley Fever"                                  
#> [27] "Vibriosis"                                     
#> [28] "Water Quality"                                 
#> [29] "Wildfire"                                      
#> [30] "Wind Damage"                                   
#> [31] "Winter Recreation"                             

### Return a dataframe of all of the sectors in FrEDI (sector names and additional information)
get_sectorInfo(description=T, gcmOnly=T)
#>                                       sector include model_type
#> 1          ATS Temperature-Related Mortality       1        GCM
#> 2                              Asphalt Roads       1        GCM
#> 3                            CIL Agriculture       1        GCM
#> 4                                  CIL Crime       1        GCM
#> 5          CIL Temperature-Related Mortality       1        GCM
#> 6      Climate-Driven Changes in Air Quality       1        GCM
#> 7              Electricity Demand and Supply       1        GCM
#> 8  Electricity Transmission and Distribution       1        GCM
#> 9                        Extreme Temperature       1        GCM
#> 10                     Forestry Surplus Loss       1        GCM
#> 11                           Inland Flooding       1        GCM
#> 12                                     Labor       1        GCM
#> 13                             Learning Loss       1        GCM
#> 14                              Lyme Disease       1        GCM
#> 15                          Marine Fisheries       1        GCM
#> 16                        Outdoor Recreation       1        GCM
#> 17                                      Rail       1        GCM
#> 18                                     Roads       1        GCM
#> 19                            Southwest Dust       1        GCM
#> 20                                   Suicide       1        GCM
#> 21                            Urban Drainage       1        GCM
#> 22                              Valley Fever       1        GCM
#> 23                                 Vibriosis       1        GCM
#> 24                             Water Quality       1        GCM
#> 25                                  Wildfire       1        GCM
#> 26                               Wind Damage       1        GCM
#> 27                         Winter Recreation       1        GCM
#>                                                               variants
#> 1              Mean, Low Confidence Interval, High Confidence Interval
#> 2                                                                  N/A
#> 3                    With CO2 Fertilization, Without CO2 Fertilization
#> 4                                                                  N/A
#> 5            Median, Low Confidence Interval, High Confidence Interval
#> 6                                       2011 Emissions, 2040 Emissions
#> 7                                                                  N/A
#> 8  No Additional Adaptation, Proactive Adaptation, Reactive Adaptation
#> 9                                 Adaptation, No Additional Adaptation
#> 10                                          SSP2 Pathway, SSP5 Pathway
#> 11                                                                 N/A
#> 12                                                                 N/A
#> 13                                 No Adaptation, Proactive Adaptation
#> 14                                                                 N/A
#> 15                                                                 N/A
#> 16                                 Proactive Adaptation, No Adaptation
#> 17 No Additional Adaptation, Proactive Adaptation, Reactive Adaptation
#> 18 No Additional Adaptation, Proactive Adaptation, Reactive Adaptation
#> 19                                                                 N/A
#> 20                                                                 N/A
#> 21                                                                 N/A
#> 22                                                                 N/A
#> 23                                                                 N/A
#> 24                                                                 N/A
#> 25                                                                 N/A
#> 26                                                                 N/A
#> 27                                                                 N/A
#>    impactYears
#> 1           NA
#> 2           NA
#> 3           NA
#> 4           NA
#> 5           NA
#> 6           NA
#> 7           NA
#> 8           NA
#> 9   2010, 2090
#> 10          NA
#> 11          NA
#> 12          NA
#> 13          NA
#> 14          NA
#> 15          NA
#> 16          NA
#> 17          NA
#> 18          NA
#> 19          NA
#> 20          NA
#> 21          NA
#> 22          NA
#> 23          NA
#> 24          NA
#> 25          NA
#> 26          NA
#> 27  2010, 2090
#>                                                                                                                                                                             impactTypes
#> 1                                                                                                                                                                                   N/A
#> 2                                                                                                                                                                                   N/A
#> 3                                                                                                                                                         Cotton, Maize, Soybean, Wheat
#> 4                                                                                                                                                                     Property, Violent
#> 5                                                                                                                                                                                   N/A
#> 6  Ozone Mortality, PM2.5 Mortality, Ozone Childhood Asthma ED Vists , Ozone New Childhood Asthma, Ozone School Loss Days, PM2.5 Childhood Asthma ED Vists , PM2.5 New Childhood Asthma
#> 7                                                                                                                                                                                   N/A
#> 8                                                                                                                                                                                   N/A
#> 9                                                                                                                                                                             Cold, Hot
#> 10                                                                                                                                                                                  N/A
#> 11                                                                                                                                                                                  N/A
#> 12                                                                                                                                                                                  N/A
#> 13                                                                                                                                               Lost Future Earnings, Adaptation Costs
#> 14                                                                                                                                                                                  N/A
#> 15                                                                                                                                                                                  N/A
#> 16                                                                                    Fishing Activity, Snow Sport Activity, Water Sport Activity, Other Activity, Other Activity Trips
#> 17                                                                                                                                                                                  N/A
#> 18                                                                                                                                                                                  N/A
#> 19                                                               Acute Myocardial Infarction, All Cardiovascular, All Mortality, All Respiratory, Asthma ER, PM2.5 New Childhood Asthma
#> 20                                                                                                                                                                                  N/A
#> 21                                                                                                                                                                                  N/A
#> 22                                                                                                                                                     Lost Wages, Morbidity, Mortality
#> 23                                                                                                                                            Direct Medical Cost, Lost Days, Mortality
#> 24                                                                                                                                                                                  N/A
#> 25                                                                                                               Other Morbidity, Mortality, Response Costs, PM2.5 New Childhood Asthma
#> 26                                                                                                                                                                                  N/A
#> 27                                                                                                                                    Alpine Skiing, Cross-Country Skiing, Snowmobiling
#>    interpolateEstimateYears
#> 1                         0
#> 2                         0
#> 3                         0
#> 4                         0
#> 5                         0
#> 6                         0
#> 7                         0
#> 8                         0
#> 9                         1
#> 10                        0
#> 11                        0
#> 12                        0
#> 13                        0
#> 14                        0
#> 15                        0
#> 16                        0
#> 17                        0
#> 18                        0
#> 19                        0
#> 20                        0
#> 21                        0
#> 22                        0
#> 23                        0
#> 24                        0
#> 25                        0
#> 26                        0
#> 27                        1

### Return a character vector with only the names of the temperature-driven sectors:
get_sectorInfo(gcmOnly=T)
#>  [1] "ATS Temperature-Related Mortality"        
#>  [2] "Asphalt Roads"                            
#>  [3] "CIL Agriculture"                          
#>  [4] "CIL Crime"                                
#>  [5] "CIL Temperature-Related Mortality"        
#>  [6] "Climate-Driven Changes in Air Quality"    
#>  [7] "Electricity Demand and Supply"            
#>  [8] "Electricity Transmission and Distribution"
#>  [9] "Extreme Temperature"                      
#> [10] "Forestry Surplus Loss"                    
#> [11] "Inland Flooding"                          
#> [12] "Labor"                                    
#> [13] "Learning Loss"                            
#> [14] "Lyme Disease"                             
#> [15] "Marine Fisheries"                         
#> [16] "Outdoor Recreation"                       
#> [17] "Rail"                                     
#> [18] "Roads"                                    
#> [19] "Southwest Dust"                           
#> [20] "Suicide"                                  
#> [21] "Urban Drainage"                           
#> [22] "Valley Fever"                             
#> [23] "Vibriosis"                                
#> [24] "Water Quality"                            
#> [25] "Wildfire"                                 
#> [26] "Wind Damage"                              
#> [27] "Winter Recreation"                        

### Run FrEDI for only the temperature-driven sectors and view results:
df_x <- run_fredi(sectorList=get_sectorInfo(gcmOnly=T))
#> Checking scenarios...
#> Calculating impacts...
#> 		Formatting initial results...
#> 		Calculating temperature-driven scaled impacts...
#> Formatting results...
#> Aggregating impacts...
#> 
#> Finished.