This function provides information about the sectors in FrEDI.

This helper function returns a character vector of sector names (default) or a data frame of sectors with related information (e.g., associated variants, impact types, etc.), which the user can supply to the run_fredi() sectorList argument.

Usage

get_sectorInfo(description = F, gcmOnly = F, slrOnly = F)

Arguments

description

= FALSE. Logical value indicating whether to include information about each sector. Returns a dataframe if description=TRUE and returns a character vector of sector names if description=FALSE (default).

gcmOnly

= FALSE. Logical value indicating whether to return only sectors with climate impacts modeled using global climate model (GCM) results.

slrOnly

= FALSE. Logical value indicating whether to return only sectors with climate impacts modeled using sea level rise (SLR) scenarios.

Value

If description=FALSE (default), outputs a character vector containing the names of sectors available for FrEDI.

If description=TRUE, `, outputs a dataframe containing the names of sectors available for FrEDI in one column, with information about the sector model type, variants, impact years, and impact types in the remaining columns.

Details

If description=FALSE (default), this helper function returns a character vector of sector names, which the user can supply to the sectorList argument to run_fredi(). If description=TRUE, get_sectorInfo() returns dataframe of sectors with related information returns a dataframe containing the sectors available for FrEDI along with additional information. Sector names are in the first column, with additional columns for the associated model type ("GCM" or "SLR"), variants, impact years, and impact types in the remaining columns. Variants, impact years, and impact types vary by sector.

Users can specify whether to return only GCM sectors or SLR sectors by setting gcmOnly=TRUE or slrOnly=TRUE, respectively. get_sectorInfo() will return the sectors in the form specified by description (see above).

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Return a character vector with the names of all of the sectors in FrEDI:
get_sectorInfo()
#>  [1] "ATS Temperature-Related Mortality"             
#>  [2] "Asphalt Roads"                                 
#>  [3] "CIL Agriculture"                               
#>  [4] "CIL Crime"                                     
#>  [5] "CIL Temperature-Related Mortality"             
#>  [6] "Climate-Driven Changes in Air Quality"         
#>  [7] "Coastal Properties"                            
#>  [8] "Electricity Demand and Supply"                 
#>  [9] "Electricity Transmission and Distribution"     
#> [10] "Extreme Temperature"                           
#> [11] "Inland Flooding"                               
#> [12] "Labor"                                         
#> [13] "Marine Fisheries"                              
#> [14] "Rail"                                          
#> [15] "Roads"                                         
#> [16] "Southwest Dust"                                
#> [17] "Suicide"                                       
#> [18] "Transportation Impacts from High Tide Flooding"
#> [19] "Urban Drainage"                                
#> [20] "Valley Fever"                                  
#> [21] "Vibriosis"                                     
#> [22] "Water Quality"                                 
#> [23] "Wildfire"                                      
#> [24] "Wind Damage"                                   
#> [25] "Winter Recreation"                             

### Return a dataframe of all of the sectors in FrEDI (sector names and additional information)
get_sectorInfo(description=T, gcmOnly=T)
#>                                       sector include model_type
#> 1          ATS Temperature-Related Mortality       1        GCM
#> 2                              Asphalt Roads       1        GCM
#> 3                            CIL Agriculture       1        GCM
#> 4                                  CIL Crime       1        GCM
#> 5          CIL Temperature-Related Mortality       1        GCM
#> 6      Climate-Driven Changes in Air Quality       1        GCM
#> 7              Electricity Demand and Supply       1        GCM
#> 8  Electricity Transmission and Distribution       1        GCM
#> 9                        Extreme Temperature       1        GCM
#> 10                           Inland Flooding       1        GCM
#> 11                                     Labor       1        GCM
#> 12                          Marine Fisheries       1        GCM
#> 13                                      Rail       1        GCM
#> 14                                     Roads       1        GCM
#> 15                            Southwest Dust       1        GCM
#> 16                                   Suicide       1        GCM
#> 17                            Urban Drainage       1        GCM
#> 18                              Valley Fever       1        GCM
#> 19                                 Vibriosis       1        GCM
#> 20                             Water Quality       1        GCM
#> 21                                  Wildfire       1        GCM
#> 22                               Wind Damage       1        GCM
#> 23                         Winter Recreation       1        GCM
#>                                                               variants
#> 1              Mean, Low Confidence Interval, High Confidence Interval
#> 2                                                                  N/A
#> 3                    With CO2 Fertilization, Without CO2 Fertilization
#> 4                                                                  N/A
#> 5            Median, Low Confidence Interval, High Confidence Interval
#> 6                                       2011 Emissions, 2040 Emissions
#> 7                                                                  N/A
#> 8  No Additional Adaptation, Proactive Adaptation, Reactive Adaptation
#> 9                                 Adaptation, No Additional Adaptation
#> 10                                                                 N/A
#> 11                                                                 N/A
#> 12                                                                 N/A
#> 13 No Additional Adaptation, Proactive Adaptation, Reactive Adaptation
#> 14 No Additional Adaptation, Proactive Adaptation, Reactive Adaptation
#> 15                                                                 N/A
#> 16                                                                 N/A
#> 17                                                                 N/A
#> 18                                                                 N/A
#> 19                                                                 N/A
#> 20                                                                 N/A
#> 21                                                                 N/A
#> 22                                                                 N/A
#> 23                                                                 N/A
#>    impactYears
#> 1           NA
#> 2           NA
#> 3           NA
#> 4           NA
#> 5           NA
#> 6           NA
#> 7           NA
#> 8           NA
#> 9   2010, 2090
#> 10          NA
#> 11          NA
#> 12          NA
#> 13          NA
#> 14          NA
#> 15  2010, 2090
#> 16          NA
#> 17          NA
#> 18          NA
#> 19          NA
#> 20          NA
#> 21          NA
#> 22          NA
#> 23  2010, 2090
#>                                                                                   impactTypes
#> 1                                                                                         N/A
#> 2                                                                                         N/A
#> 3                                                               Cotton, Maize, Soybean, Wheat
#> 4                                                                           Property, Violent
#> 5                                                                                         N/A
#> 6                                                                                Ozone, PM2.5
#> 7                                                                                         N/A
#> 8                                                                                         N/A
#> 9                                                                                   Cold, Hot
#> 10                                                                                        N/A
#> 11                                                                                        N/A
#> 12                                                                                        N/A
#> 13                                                                                        N/A
#> 14                                                                                        N/A
#> 15 Acute Myocardial Infarction, All Cardiovascular, All Mortality, All Respiratory, Asthma ER
#> 16                                                                                        N/A
#> 17                                                                                        N/A
#> 18                                                           Lost Wages, Morbidity, Mortality
#> 19                                                  Direct Medical Cost, Lost Days, Mortality
#> 20                                                                                        N/A
#> 21                                                       Morbidity, Mortality, Response Costs
#> 22                                                                                        N/A
#> 23                                          Alpine Skiing, Cross-Country Skiing, Snowmobiling

### Return a character vector with only the names of the temperature-driven sectors:
get_sectorInfo(gcmOnly=T)
#>  [1] "ATS Temperature-Related Mortality"        
#>  [2] "Asphalt Roads"                            
#>  [3] "CIL Agriculture"                          
#>  [4] "CIL Crime"                                
#>  [5] "CIL Temperature-Related Mortality"        
#>  [6] "Climate-Driven Changes in Air Quality"    
#>  [7] "Electricity Demand and Supply"            
#>  [8] "Electricity Transmission and Distribution"
#>  [9] "Extreme Temperature"                      
#> [10] "Inland Flooding"                          
#> [11] "Labor"                                    
#> [12] "Marine Fisheries"                         
#> [13] "Rail"                                     
#> [14] "Roads"                                    
#> [15] "Southwest Dust"                           
#> [16] "Suicide"                                  
#> [17] "Urban Drainage"                           
#> [18] "Valley Fever"                             
#> [19] "Vibriosis"                                
#> [20] "Water Quality"                            
#> [21] "Wildfire"                                 
#> [22] "Wind Damage"                              
#> [23] "Winter Recreation"                        

### Run FrEDI for only the temperature-driven sectors and view results:
df_x <- run_fredi(sectorList=get_sectorInfo(gcmOnly=T))
#> Checking scenarios...
#> Calculating impacts...
#> 		Formatting initial results...
#> 		Calculating temperature-driven scaled impacts...
#> Formatting results...
#> Aggregating impacts...
#> 
#> Finished.