Retrieve a character vector of sectors available in the FrEDI SV module (run_fredi_sv) or a data frame with SV sectors and additional information.

get_sv_sectorInfo returns a character vector with the names of sectors in the FrEDI SV module (default) or a data frame of SV sectors with additional information (e.g., associated variants, model type, etc.).

Usage

get_sv_sectorInfo(description = F, gcmOnly = F, slrOnly = F)

Arguments

description

Logical value indicating whether to include information about each sector. Returns a data frame if description=TRUE and returns a character vector of sector names if description=FALSE (default).

gcmOnly

Logical value indicating whether to return only sectors with impacts driven by temperature changes.

slrOnly

Logical value indicating whether to return only sectors with impacts driven by sea level changes.

Value

get_sv_sectorInfo() will return SV sectors in the form specified by description:

If description=FALSE (default), outputs a character vector containing the names of sectors available for the FrEDI SV Module.

If description=TRUE, outputs a data frame containing the names of sectors available for the FrEDI SV Module in one column, with information about the sector model type, model type ("GCM" or "SLR") and associated driver unit ("degrees Celsius" or "cm", respectively ), impact units (e.g., mortality, etc.), and available variants in the remaining columns.

Details

get_sv_sectorInfo returns a character vector with the names of sectors in the FrEDI SV module (description=FALSE, default) or a data frame with SV sectors and additional information (description=TRUE).

If description=FALSE (default), get_sv_sectorInfo returns a character vector of names of sectors that can be passed to the run_sv_fredi() sector argument. Specify whether to return only GCM sectors by running get_sv_sectorInfo(gcmOnly=TRUE) or SLR sectors by running get_sv_sectorInfo(slrOnly=TRUE).

If description=TRUE, get_sv_sectorInfo returns a data frame of sectors with related information, such as whether a particular sector is driven primarily by temperature (modelType="GCM") or sea level rise (modelType="SLR"), associated driver units ("degrees Celsius" for temperature-driven sectors, "cm" for SLR-driven sectors), impact units (e.g., mortality, etc.), and sector variants. Users can use get_sv_sectorInfo to determine which sectors can be passed to the run_sv_fredi() sector argument and/or to determine whether a particular sector is driven primarily by temperature (modelType="GCM") or sea level rise (modelType="SLR").

Users can specify whether to return only GCM sectors by setting gcmOnly=TRUE or SLR sectors by setting slrOnly=TRUE.

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Return a character vector with the names of all of the sectors in the FrEDI SV Module:
get_sv_sectorInfo()
#> Error in get_sv_sectorInfo(): object 'svDataList' not found

### Return a data frame of all of the sectors in the FrEDI SV Module (sector names and additional information)
get_sv_sectorInfo(description=T)
#> Error in get_sv_sectorInfo(description = T): object 'svDataList' not found

### Return a character vector with only the names of the temperature-driven sectors:
get_sv_sectorInfo(gcmOnly=T)
#> Error in get_sv_sectorInfo(gcmOnly = T): object 'svDataList' not found

### Return a character vector with only the names of the temperature-driven sectors:
get_sv_sectorInfo(slrOnly=T)
#> Error in get_sv_sectorInfo(slrOnly = T): object 'svDataList' not found