Import custom scenarios for climate and socioeconomics (temperature and global mean sea level rise (GMSL), population, and GDP) from user-specified file names

This function enables users to import data on custom scenarios for use with FREDI (supplied as inputs to run_fredi()). Users specify path names to CSV files containing temperature, global mean sea level rise (GMSL), gross domestic product (GDP), and population scenarios. import_inputs() reads in and format any specified files as data frames and returns a list of data frames for imported scenarios.

Usage

import_inputs(
  tempfile = NULL,
  slrfile = NULL,
  popfile = NULL,
  gdpfile = NULL,
  temptype = "conus",
  popform = "wide"
)

Arguments

tempfile

A character string indicating the location of a CSV file containing a custom temperature scenario (first column contains years; second column contains temperatures, in degrees Celsius, above the 1995 baseline year). The temperature scenario must start in 2000 or earlier and end at or after the maximum model run year (e.g., as specified by the maxYear argument to run_fredi()).

slrfile

A character string indicating the location of a CSV file containing a custom sea level rise scenario (first column contains years; second column contains values for global mean sea level rise (GMSL), in centimeters, above the 2000 baseline). The SLR scenario must start in 2000 or earlier and end at or after the maximum model run year (e.g., as specified by the maxYear argument to run_fredi()).

popfile

A character string indicating the location of a CSV file containing a custom population scenario for NCA regions. The first column contains years in the interval 2010 to 2300. The number of additional columns, column names, and column contents depend on the population format set by popform. For more details, see popform. The population scenario must start in 2010 or earlier and end at or after the maximum model run year (e.g., as specified by the maxYear argument to run_fredi()).

gdpfile

A character string indicating the location of a CSV file containing a custom scenario for gross domestic product (GDP) (first column contains years; second column contains values for GDP, in total 2015$). The GDP scenario must start in 2010 or earlier and end at or after the maximum model run year (e.g., as specified by the maxYear argument to run_fredi()).

temptype

A character string indicating whether the temperature values in the temperature input file (specified by tempfile represent global temperature change (temptype="global") or temperature change for the contiguous U.S. (temptype="conus") in degrees Celsius. By default, the model assumes temperatures are CONUS temperatures (i.e., temptype="conus").

popform

A character string indicating whether the populations in the population input file specified by popfile are spread across multiple columns (i.e., popform="wide") or are combined in a single column (i.e., popform="long"). By default, the model assumes popform="wide". For both formats (popform="wide" or popform="long"), the first column contains years. If popform="wide" (default), the second through eighth columns of popfile must contain population values for each NCA region, with the associated NCA region as the column name. If popform="long", the second column must contain NCA region names and the third column must contain values for the associated region population.

Value

import_inputs() returns a list of named elements containing data frames with custom scenarios for temperature, GMSL, GDP, and regional population, respectively:

List Index	Description

tempInput	Data frame containing a custom temperature scenario imported from the CSV file specified by tempfile, with missing values removed. tempInput has two columns with names c("year", "temp_C") containing the year and CONUS temperatures in degrees Celsius, respectively.

slrInput	Data frame containing a custom GMSL scenario imported from the CSV file specified by slrfile, with missing values removed. slrInput has two columns with names c("year", "slr_cm") containing the year and global mean sea level rise (GMSL) in centimeters, respectively.

gdpInput	Data frame containing a custom GDP scenario imported from the CSV file specified by gdpfile, with missing values removed. gdpInput has two columns with names c("year", "gdp_usd") containing the year and the U.S. national GDP in 2015$, respectively.

popInput	Data frame containing a custom temperature scenario imported from the CSV file specified by popfile, with missing values removed. popInput has and three columns with names c("year", "region", "reg_pop") containing the year, the NCA region name, and the NCA region population, respectively.

Details

This function enables users to import data on custom scenarios for use with temperature binning. Users specify path names to CSV files containing temperature, global mean sea level rise (GMSL), population, and gross domestic product (GDP) scenarios (tempfile, slrfile, gdpfile, and popfile, respectively). import_inputs() reads in and formats any specified files as data frames and returns a list of data frames for imported scenarios. Users can specify whether the temperature input is for the contiguous U.S. (CONUS) or global using temptype and specify the format of the population scenario using popform.

Values for input scenarios must be within reasonable ranges. Temperature and GMSL inputs must begin in 2000 or earlier, whereas values for population and GDP scenarios can start in 2010 or earlier.

The input temperature scenario (passed to run_fredi() via the inputsList argument) requires temperatures for CONUS in degrees Celsius relative to 1995 (degrees of warming relative to the baseline). Temperature values must be greater than or equal to zero degrees if warming (CONUS). Users can convert global temperatures to CONUS temperatures using convertTemps(from="global") or by specifying import_inputs(temptype="global") when importing a temperature scenario from a CSV file.

Values for the sea level rise (SLR) scenario are for global mean sea level rise (GMSL) must be in centimeters (cm) and values must be greater than or equal to zero and less than or equal to 250 cm.

Population and gross domestic product (GDP) values must be greater than or equal to zero.

If the population input is spread across multiple columns (i.e., popform="wide"), columns must be named according to the NCA regions. If the population input is in the long format, the region value must be in the second column. The NCA region names for population inputs must be in the following character vector: c("Midwest", "Northeast", "Northern.Plains", "Northwest", "Southeast", "Southern.Plains", "Southwest"). All regions must be present in the population input file.

import_inputs() outputs a list of data frames that can be passed to the main FREDI function run_fredi() using the inputList argument. For example, specify run_fredi(inputsList=x) to generate impacts for a custom scenario x (where x is a list of data frames such as that output from import_inputs()) (see run_fredi()).

All inputs to import_inputs() are optional. If the user does not specify a file path for tempfile, slrfile, gdpfile, or popfile (or if there is an error reading in inputs from those file paths), import_inputs() outputs a list with a NULL value for the associated list element. run_fredi() defaults back to the default scenarios for any elements of the inputs list that are NULL or missing. In other words, running run_fredi(inputsList=list()) returns the same outputs as running run_fredi() (see run_fredi()).

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Path to example scenarios
scenariosPath <- system.file(package="FrEDI") %>% file.path("extdata","scenarios")

### View example scenario names
scenariosPath %>% list.files
#> [1] "GCAM_scenario.csv" "pop_scenario.csv"  "slr_from_GCAM.csv"

### Temperature Scenario File Name
tempInputFile <- scenariosPath %>% file.path("GCAM_scenario.csv")

### SLR Scenario File Name
slrInputFile  <- scenariosPath %>% file.path("slr_from_gcam.csv")

### Population Scenario File Name
popInputFile  <- scenariosPath %>% file.path("pop_scenario.csv")

### Import inputs
example_inputsList <- import_inputs(
  tempfile = tempInputFile,
  slrfile  = slrInputFile,
  popfile  = popInputFile,
  temptype = "global",
  popform  = "wide"
)
#> 
#> In import_inputs():
#> 	User supplied temperature input...
#> 		Importing data from C:/Users/emcduf01/OneDrive - Environmental Protection Agency (EPA)/Documents/R/R-4.0.3/library/FrEDI/extdata/scenarios/GCAM_scenario.csv...
#> 		Formatting temperature inputs...
#> 			User specified `temptype='global'`...
#> 			Converting global temperatures to CONUS temperatures...
#> 	User supplied SLR input...
#> 		Importing data from C:/Users/emcduf01/OneDrive - Environmental Protection Agency (EPA)/Documents/R/R-4.0.3/library/FrEDI/extdata/scenarios/slr_from_gcam.csv...
#> 		Formatting SLR inputs...
#> 	User supplied regional population input...
#> 		Importing data from C:/Users/emcduf01/OneDrive - Environmental Protection Agency (EPA)/Documents/R/R-4.0.3/library/FrEDI/extdata/scenarios/pop_scenario.csv...
#> 		Formatting regional population inputs...
#> 			User specified `popform='wide'`...Gathering population by region...
#> Finished.

### Use imports with FREDI:
df_x <- run_fredi(inputsList=example_inputsList)
#> Checking input values...
#> Creating temperature scenario from user inputs...
#> Creating SLR scenario from user inputs...
#> No GDP scenario provided...Using default GDP scenario...
#> Creating Population scenario from user inputs...
#> Formatting results...
#> 	Interpolating between impact year estimates...
#> 	Calculating model averages...
#> 	Calculating national totals...
#> 	Summing across impact types...
#> 
#> Finished.