Project annual average impacts from temperature and sea level change throughout the 21st century for available sectors

This function allows users to project annual average impacts from temperature and sea level change through 2100 (2010-2100) for available sectors, with the option to extend results to 2300 (2010-2300). Users may specify custom temperature, sea level, U.S. population, and GDP scenarios. The output is an R data frame object containing annual average impacts, by year, for each sector, variant, impact type, region, state, and model.

As of FrEDI Version 4.1.0, run_fredi() calculates impacts at the state-level for all available sectors.

Usage

run_fredi(
  inputsList = list(temp = NULL, slr = NULL, gdp = NULL, pop = NULL),
  sectorList = NULL,
  aggLevels = c("national", "modelaverage", "impactyear", "impacttype"),
  elasticity = 1,
  maxYear = 2100,
  thru2300 = FALSE,
  outputList = FALSE,
  allCols = FALSE,
  silent = TRUE
)

Arguments

inputsList=list(gdp=NULL, pop=NULL, temp=NULL, slr=NULL)

A list with named elements (gdp, pop, temp, and/or slr), each containing data frames of custom scenarios for gross domestic product (GDP), state-level population, temperature, and/or global mean sea level rise (GMSL) trajectories, respectively, over a continuous period. Temperature and sea level rise inputs should start in 2000 or earlier. Values for population and GDP scenarios can start in 2010 or earlier. Values for each scenario type must be within reasonable ranges. For more information, see import_inputs().

sectorList=NULL

A character vector indicating a selection of sectors for which to calculate results (see get_sectorInfo()). If NULL, all sectors are included (i.e., sectorList = get_sectorInfo()).

aggLevels="all"

Levels of aggregation at which to summarize data: one or more of c("national", "modelaverage", "impactyear", "impacttype", "all", "none"). Defaults to all levels (i.e., aggLevels = "all"). Uses the same aggregation levels as aggregate_impacts(). Note that, if "impacttype" is in aggLevels (e.g., aggLevels = "all"), columns "physical_measure" and "physical_impacts" will be dropped from the results data frame. This is because aggregating over impact types for some sectors requires summing costs over different types of physical impacts, so reporting the physical impacts would be nonsensical.

elasticity=1

A numeric value indicating an elasticity to use for adjusting VSL for applicable sectors and impacts (defaults to elasticity = 1). Applicable sectors and impacts are: Air Quality (all impact types), ATS Temperature-Related Mortality (impactType = "N/A"; i.e., all impact types), CIL Temperature-Related Mortality, Extreme Temperature (all impact types), Suicide (impactType = "N/A"; i.e., all impact types), Southwest Dust (impactType = "All Mortality"), Valley Fever (impactType = "Mortality"), Vibriosis (impactType = "N/A"; i.e., all impact types), and Wildfire (impactType = "Mortality").

maxYear=2100

A numeric value indicating the maximum year for the analysis. The range for maxYear is [2011, 2300]. Defaults to maxYear = 2100.

thru2300=FALSE

A TRUE/FALSE shortcut that overrides the maxYear argument to run the model to 2300. Defaults to thru2300 = FALSE.

outputList=FALSE

A TRUE/FALSE value indicating whether to output results as a data frame object (outputList = FALSE, default) or to return a list of objects (outputList = TRUE) that includes information about model provenance (including input arguments and input scenarios) along with the data frame of results.

allCols=FALSE

A TRUE/FALSE value indicating whether to include intermediate column values in results (e.g., physical and economic multipliers). Used in testing. Note that aggregation levels must be set to aggLevels = "none" to properly return the intermediate columns. Defaults to allCols = FALSE).

silent=TRUE

A TRUE/FALSE value indicating the level of messaging desired by the user (default=TRUE).

Value

If outputList=FALSE, the output of run_fredi() is a data frame object (described above) containing annual average impacts over the analysis period, for each sector, variant, impact type, impact year, region, state, and model (GCM name for temperature-driven sectors and "Interpolation" for SLR-driven sectors).

If outputList=TRUE, run_fredi() returns a list object containing the following:

statusList. A list with values for the arguments passed to run_fredi() (including defaults if unspecified).

argsList. A list with elements named after run_fredi() arguments, containing the values of the arguments passed to run_fredi() (or default values if unspecified).

scenarios. A list with named elements "temp", "slr", "gdp", and "pop" -- each containing the scenarios for temperature, SLR, GDP, and population as used by the model in calculating impacts.

results. Containing a data frame of annual impacts (i.e., the same data frame returned if outputList = FALSE).

Details

This function allows users to project annual average impacts from temperature and sea level change through 2300 (2010-2300) for available sectors. run_fredi() is the main function in the FrEDI R package, described elsewhere (See https://epa.gov/cira/FrEDI for more information).

Users can specify an optional list of custom scenarios with inputsList (for more information on the format of inputs, see import_inputs()). The function import_inputs() can be used to importing custom scenarios from CSV files. import_inputs() returns a list with named elements gdp, pop, temp, and slr, with each containing a data frame with a custom scenario for GDP, state-level population, temperature, and/or GMSL respectively. If a user imports scenarios using import_inputs(), they can pass the outputs of import_inputs() directly to the run_fredi() argument inputsList. Note that the documentation for import_inputs() can also provide additional guidance and specification on the formats for each scenario type.

Otherwise, run_fredi() looks for a list object passed to the argument inputsList. Within that list, run_fredi() looks for named list elements -- gdp, pop, temp, and/or slr -- each containing a data frame with a custom scenario for GDP, state population, temperature, and/or GMSL, respectively. If inputsList = NULL or inputsList = list() (default), run_fredi() uses default trajectories for GDP, population, temperature, and SLR (see gdpDefault, popDefault, and gcamScenarios for more information). run_fredi() will default back to the default scenarios for any list elements that empty or NULL (in other words, running run_fredi(inputsList = list()) returns the same outputs as running run_fredi()).

GDP Inputs. The input scenario for gross domestic product (GDP) requires national GDP values in 2015$. GDP values must be greater than or equal to zero.

gdp requires a data frame object with two columns with names "year", and "gdp_usd" containing the year and the national GDP, respectively. GDP values must be greater than or equal to zero.

GDP inputs must have at least one non-missing value in 2010 or earlier and at least one non-missing value in or after the final analysis year (as specified by maxYear).

If the user does not specify an input scenario for GDP (i.e., inputsList = list(gdp = NULL), run_fredi() uses a default GDP scenario.

Population Inputs. The input population scenario requires state-level population values. Population values must be greater than or equal to zero.

pop requires a data frame object with five columns with names "region", "state", "postal", "year", and "pop" containing the year, the NCA region name (one of "Midwest", "Northeast", "Northern Plains", "Northwest", "Southeast", "Southern Plains", or "Southwest"), the state name, the postal code abbreviation for the state, and the state population, respectively.

Population inputs must have at least one non-missing value in 2010 or earlier and at least one non-missing value in or after the final analysis year (as specified by maxYear).

If the user does not specify an input scenario for population (i.e., inputsList = list(pop = NULL), run_fredi() uses a default population scenario.

Temperature Inputs. The input temperature scenario requires CONUS temperatures in degrees Celsius relative to 1995 (degrees of warming relative to the baseline year--i.e., the central year of the 1986-2005 baseline). CONUS temperature values must be greater than or equal to zero degrees Celsius.

Users can convert global temperatures to CONUS temperatures using convertTemps(from = "global") (or by specifying import_inputs(temptype = "global") when using import_inputs() to import a temperature scenario from a CSV file).

temp requires a data frame object with two columns with names "year", and "temp_C" containing the year and CONUS temperatures in degrees Celsius, respectively.

Temperature inputs must have at least one non-missing value in 2000 or earlier and at least one non-missing value in or after the final analysis year (as specified by maxYear).

If the user does not specify an input scenario for temperature (i.e., inputsList = list(temp = NULL), run_fredi() uses a default temperature scenario.

SLR Inputs. The input SLR scenario requires values for changes in global mean sea level rise (GMSL) heights in centimeters (cm). GMSL heights must be greater than or equal to zero.

slr requires a data frame object with two columns with names "year", "slr_cm" containing the year and global mean sea level rise (GMSL) in centimeters, respectively.

SLR inputs must have at least one non-missing value in 2000 or earlier and at least one non-missing value in or after the final analysis year (as specified by maxYear).

If the user does not specify an input scenario for SLR (i.e., inputsList = list(slr = NULL), run_fredi() first converts the input or default CONUS temperature scenario to global temperatures (using convertTemps()) and then converts the global temperatures to a global mean sea level rise (GMSL) height in centimeters (using temps2slr()).

run_fredi() linearly interpolates missing annual values for all input scenarios using non-missing values (each scenario requires at least two non-missing values as detailed above for each scenario type). After interpolation of the input scenarios, run_fredi() subsets the input scenarios to values within the analysis period.

Temperatures are interpolated using 1995 as the baseline year (i.e., the central year of the 1986-2005 baseline) and GMSL is interpolated using 2000 as the baseline year. In other words, temperature (in degrees Celsius) is set to zero for the year 1995, whereas GMSL is set to zero for the year 2000. The interpolated temperature and GMSL scenarios are combined into a column called driverValue, along with additional columns for year, the driver unit (column "driverUnit", with driverUnit = "degrees Celsius" and driverUnit = "cm" for temperature- and SLR-driven sectors, respectively), and the associated model type (column "model_type", with model_type = "GCM" and model_type = "SLR" for temperature- and SLR-driven sectors, respectively.

run_fredi() calculates national population from state-level values and then calculates GDP per capita from values for GDP and national population. Values for state population, national population, national GDP (in 2015$), and national per capita GDP (in 2015$/capita) are provided in the results data frame in columns "pop", "national_pop", "gdp_usd", and "gdp_percap", respectively.

By default, run_fredi() will calculate impacts for all sectors included in the tool. Alternatively, users can pass a character vector specifying a single sector or a subset of sectors using the sectorList argument. To see a list of sectors included within FrEDI, run get_sectorInfo(). If sectorList = NULL (default), all sectors are included.

By default, run_fredi() calculates impacts starting in the year 2010 and ending in 2100. Specify an alternative end year for the analysis using the maxYear argument. maxYear has a default value of 2100 and minimum and maximum values of 2011 and 2300, respectively. Alternatively, users can set argument thru2300 = TRUE to override the maxYear argument and set maxYear = 2300. Note that the default scenarios included within FrEDI stop in the year 2100; users must provide custom input scenarios out to the desired end year and specify a maxYear >= 2100 (and maxYear <= 2300) in order to return non-missing values for years after 2100.

Annual impacts for each sector, variant, impact type, and impact year combination included in the model are calculated by multiplying scaled impacts by a physical scalar and economic scalars and multipliers. Some sectors use Value of a Statistical Life (VSL) to adjust the value non-linearly over time. run_fredi() uses a default value of elasticity = 1to adjust VSL for applicable sectors and impacts (the default value of elasticity = 1 keeps VSL constant over time). A custom elasticity can be passed to the elasticity argument. Applicable sectors and impacts are *Air Quality (all impact types), ATS Temperature-Related Mortality (impactType = "N/A"; i.e., all impact types), CIL Temperature-Related Mortality, Extreme Temperature (all impact types), Suicide (impactType = "N/A"; i.e., all impact types), Southwest Dust (impactType = "All Mortality"), Valley Fever (impactType = "Mortality"), Vibriosis (impactType = "N/A"; i.e., all impact types), and Wildfire (impactType = "Mortality").

run_fredi() aggregates or summarizes results to level(s) of aggregation specified by the user (passed to aggLevels) using the post-processing helper function aggregate_impacts(). Users can specify all aggregation levels at once by specifying aggLevels = "all" (default) or no aggregation levels (aggLevels = "none"). Users can specify a single aggregation level or multiple aggregation levels by passing a single character string or character vector to aggLevels. Options for aggregation include calculating national totals (aggLevels = "national"), averaging across model types and models (aggLevels = "modelaverage"), summing over all impact types (aggLevels = "impacttype"), and interpolating between impact year estimates (aggLevels = "impactYear").

If the user specifies aggLevels = "none", run_fredi() returns a data frame with columns: "sector", "variant", "impactType", "impactYear", "region", "state", "postal", "model_type", "model", "sectorprimary", "includeaggregate", "physicalmeasure", "driverType", "driverUnit", "driverValue", "gdp_usd", "national_pop", "gdp_percap", "pop", "year", "physical_impacts", and "annual_impacts".

Columns "sector", "variant", "impactType", "impactYear", "region", "state", "postal", "model_type", and "model" all contain observation identifiers (sector name, variant, impact type, impact year, region, state, state postal code abbreviation, model type, and model, respectively).

Columns "sectorprimary" and "includeaggregate" contain values that provide information about how to treat sectors and when aggregating over sectors (e.g., summing impacts across sectors). Note that FrEDI does not currently provide functionality to aggregate over sectors; this information is provided for user convenience.

Column "sectorprimary" contains values indicating which variant to use as the primary one for each sector: sectorprimary = 1for primary variants and sectorprimary = 0 for non-primary variants. When aggregating impacts over sectors, users should filter the outputs of run_fredi() to variants with sectorprimary == 1.

Column "includeaggregate" contains values that provide information about how to treat sectors when aggregating over sectors (e.g., summing impacts across sectors). Sectors that have a value of includeaggregate == 0 should be dropped when aggregating results over sectors, to avoid potential double-counting of impacts for similar sectors. For instance, sectors ATS Temperature-Related Mortality, CIL Temperature-Related Mortality, and Extreme Temperature have values for temperature-related mortality. To avoid double counting, outputs of run_fredi() should be filtered to values for which includeaggregate > 0. Sectors with a value of includeaggregate > 0 can be included when aggregating over sectors; most sectors with a value of includeaggregate > 0 will have a value of includeaggregate = 1. Values of includeaggregate > 1 flag additional information about sectors; currently, only the Suicide sector has a value of includeaggregate > 1, with a value of includeaggregate = 2. This flag indicates that the impacts from Suicide can be included when summing values across sectors, but may have some overlap with impacts from ATS Temperature-Related Mortality. For more information about the potential overlap between impacts for ATS Temperature-Related Mortality and Suicide, visit the technical documentation at https://epa.gov/cira/FrEDI/,

Columns "driverType", "driverUnit", and "driverValue" contain information about the temperature and SLR scenarios.

Columns "gdp_usd", "national_pop", "gdp_percap", and "pop" contain information about the GDP and population scenarios.

Columns "physicalmeasure" and "physical_impacts" contain information about physical impacts.

Column "annual_impacts" contains information on the economic value associated with annual impacts.

If the user specifies aggLevels = "all" or other combinations of aggregation levels, run_fredi() passes the results data frame and the aggLevels argument to the aggregate_impacts() function. aggregate_impacts() then performs the following calculations, using the default grouping columns for the aggregate_impacts(): "sector", "variant", "impactType", "impactYear", "region", "state", "postal", "model_type", "model", "sectorprimary", "includeaggregate", "physicalmeasure", and "year" (note that the "variant" column referred to below contains information about the variant name (or “N/A”), as applicable).

Aggregation Level	Description

impactyear	To aggregate over impact years, aggregate_impacts() first separates results for sectors with only one impact year estimate (i.e., impactYear = "N/A") from from observations with multiple impact year estimates (i.e., sectors with results for both impactYear = "2010" and impactYear = "2090"). For these sectors with multiple impact years, physical impacts and annual costs (columns "physical_impacts" and "annual_impacts") are linearly interpolated between impact year estimates. For any model run years above 2090, annual results for sectors with multiple impact years return the 2090 estimate. The interpolated values are then row-bound to the results for sectors with a single impact year estimate, and column impactYear set to impactYear = "Interpolation" for all values. If "impactyear" is included in aggLevels (e.g., aggLevels = "all"), aggregate_impacts() aggregates over impact years before performing other types of aggregation.

modelaverage	To aggregate over models for temperature-driven sectors, aggregate_impacts() averages physical impacts and annual costs (columns "physical_impacts" and "annual_impacts", respectively) across all GCM models present in the data. aggregate_impacts() drops the column "model" from the grouping columns when averaging over models. Averages exclude observations with missing values. However, If all values within a grouping are missing, the model average is set to NA. The values in column "model" are set to "Average" for model averages and the model averages data frame is then row-bound to the main results data frame. For SLR-driven sectors, there is no need for additional model aggregation; these values already have model = "Interpolation". If "modelaverage" is included in aggLevels (e.g., aggLevels = "all"), aggregate_impacts() first aggregates over impact years (if "impactyear" present in aggLevels or if aggLevels = "all") before aggregating over models.

national	To aggregate values to the national level, aggregate_impacts() sums physical impacts and annual costs (columns "physical_impacts" and "annual_impacts", respectively) across all states present in the data. aggregate_impacts() drops the columns "region", "state", and "postal" when summing over states and regions. Years which have missing column data for all regions return as NA. Values for columns "region", "state", and "postal" are set to "National Total", All, and US, respectively. The data frame with national totals is then row-bound to the main results data frame. If "national" is included in aggLevels (e.g., aggLevels = "all"), aggregate_impacts() first aggregates over impact years and/or models (if "impactyear" and/or "modelaverage" are present in aggLevels or if aggLevels = "all") before aggregating over models.

impacttype	To aggregate values over impact types, aggregate_impacts() sums annual impacts (column "annual_impacts") across all impact types for each sector. aggregate_impacts() drops the column "impactType" and "physicalmeasure" from the grouping columns when summing over impact types. Years which have missing column data for all impact types return as NA. All values in column "impactType" are set to "all". Aggregating over impact types, drops columns related to physical impacts (i.e., columns "physicalmeasure" and "physical_impacts"). These columns are dropped since aggregating over impact types for some sectors requires summing costs over different types of physical impacts, so reporting the physical impacts would be nonsensical.

After aggregating values, aggregate_impacts() joins the data frame of impacts with information about "driverType", "driverUnit", "driverValue", "gdp_usd", "national_pop", "gdp_percap", and "pop".

If outputList = FALSE (default), run_fredi() returns a data frame of annual average impacts over the analysis period, for each sector, variant, impact type, impact year, region, state, model type ("GCM" or "SLR"), and model. If outputList = TRUE, in addition to the data frame of impacts, run_fredi() returns a list object containing information about values for function arguments, driver scenarios, and population and GDP scenarios.

References

Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at https://epa.gov/cira/FrEDI/.

Examples

### Load FrEDI
require(FrEDI)

### Run function with defaults
run1 <- run_fredi()
#> Checking scenarios...
#> Calculating impacts...
#> 		Formatting initial results...
#> 		Calculating temperature-driven scaled impacts...
#> 		Calculating SLR-driven scaled impacts...
#> Formatting results...
#> Aggregating impacts...
#> 
#> Finished.

### Load temperature scenarios and glimpse data
data("gcamScenarios")
gcamScenarios |> glimpse()
#> Rows: 606
#> Columns: 6
#> $ year          <int> 2000, 2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 20…
#> $ temp_C_conus  <dbl> 0.1804171, 0.2238804, 0.2640659, 0.2986389, 0.3277609, 0…
#> $ temp_C_global <dbl> 0.1269649, 0.1575513, 0.1858310, 0.2101611, 0.2306551, 0…
#> $ slr_cm        <dbl> 0.0000000, 0.2727356, 0.5552278, 0.8458496, 1.1430287, 1…
#> $ scenario      <chr> "ECS_3.0_REF", "ECS_3.0_REF", "ECS_3.0_REF", "ECS_3.0_RE…
#> $ model         <chr> "Hector_GCAM_v5.3", "Hector_GCAM_v5.3", "Hector_GCAM_v5.…

### Load population scenario and glimpse data
data("popDefault")
popDefault |> glimpse()
#> Rows: 14,259
#> Columns: 5
#> $ region    <chr> "Southeast", "Southeast", "Southeast", "Southeast", "Southea…
#> $ state     <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Alabama", "Alab…
#> $ postal    <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", …
#> $ year      <int> 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, …
#> $ state_pop <dbl> 4779736, 4823623, 4867509, 4911396, 4955282, 4999169, 503604…

### Subset temperature scenario
temps1 <- gcamScenarios |> filter(scenario=="ECS_3.0_ref")
temps1 <- temps1 |> select(year, temp_C_conus)

### Run custom scenario
run2 <- run_fredi(inputsList=list(temp=temps1, pop=popDefault))
#> Checking scenarios...
#> 
#> 		Checking input values for pop inputs...
#> 			Column "pop" not found in popfile data!
#> 			Looking for columns with matches to the string "pop"...1 match found!
#> 			Using column "", and renaming to "pop"...
#> 				Data is missing the following required columns: state_pop!
#> 				Dropping pop inputs from outputs.
#> Calculating impacts...
#> 		Formatting initial results...
#> 		Calculating temperature-driven scaled impacts...
#> 		Calculating SLR-driven scaled impacts...
#> Formatting results...
#> Aggregating impacts...
#> 
#> Finished.

### Load scenarios from file:
scenariosPath <- system.file(package="FrEDI") |> file.path("extdata","scenarios")
scenariosPath |> list.files()
#> [1] "gcamDefault.csv"   "gcamScenarios.csv" "gdpDefault.csv"   
#> [4] "popDefault.csv"   

### SLR Scenario File Name
slrInputFile  <- scenariosPath |> file.path("gcamDefault.csv")

### Population Scenario File Name
popInputFile  <- scenariosPath |> file.path("popDefault.csv")

### Import inputs
x_inputs <- import_inputs(inputsList=list(slr=slrInputFile, pop=popInputFile), popArea="state")
#> 
#> In import_inputs():
#> 	Loading data...
#> 	User specified slrfile...
#> 		Importing data from C:/Users/knoiva/AppData/Local/R/win-library/4.4/FrEDI/extdata/scenarios/gcamDefault.csv...
#> 		Data loaded.
#> 	User specified popfile...
#> 		Importing data from C:/Users/knoiva/AppData/Local/R/win-library/4.4/FrEDI/extdata/scenarios/popDefault.csv...
#> 		Data loaded.
#> 
#> 	Checking input values...
#> 
#> 		Checking input values for slr inputs...
#> 		Values passed.
#> 
#> 		Checking input values for pop inputs...
#> 			Checking that all states, etc. are present...
#> 
#> 			Checking inputs for unique regions, states...
#> 				All regions present...
#> 				All states present...
#> 		Values passed.
#> 
#> Finished.

### Run custom scenarios
run3 <- run_fredi(inputsList=x_inputs)
#> Checking scenarios...
#> 
#> 		Checking input values for slr inputs...
#> 		Values passed.
#> 
#> 		Checking input values for pop inputs...
#> 			Checking that all states, etc. are present...
#> 
#> 			Checking inputs for unique regions, states...
#> 				All regions present...
#> 				All states present...
#> 		Values passed.
#> 	Creating SLR scenario from user inputs...
#> 	Creating population scenario from user inputs...
#> Calculating impacts...
#> 		Formatting initial results...
#> 		Calculating temperature-driven scaled impacts...
#> 		Calculating SLR-driven scaled impacts...
#> Formatting results...
#> Aggregating impacts...
#> 
#> Finished.

### Get information on sectors:
get_sectorInfo()
#>  [1] "ATS Temperature-Related Mortality"             
#>  [2] "Asphalt Roads"                                 
#>  [3] "CIL Agriculture"                               
#>  [4] "CIL Crime"                                     
#>  [5] "CIL Temperature-Related Mortality"             
#>  [6] "Climate-Driven Changes in Air Quality"         
#>  [7] "Coastal Properties"                            
#>  [8] "Electricity Demand and Supply"                 
#>  [9] "Electricity Transmission and Distribution"     
#> [10] "Extreme Temperature"                           
#> [11] "Inland Flooding"                               
#> [12] "Labor"                                         
#> [13] "Marine Fisheries"                              
#> [14] "Rail"                                          
#> [15] "Roads"                                         
#> [16] "Southwest Dust"                                
#> [17] "Suicide"                                       
#> [18] "Transportation Impacts from High Tide Flooding"
#> [19] "Urban Drainage"                                
#> [20] "Valley Fever"                                  
#> [21] "Vibriosis"                                     
#> [22] "Water Quality"                                 
#> [23] "Wildfire"                                      
#> [24] "Wind Damage"                                   
#> [25] "Winter Recreation"                             

### Run for a single sector, with default inputs, no aggregation, and elasticity=1:
run4 <- run_fredi(sectorList="ATS Temperature-Related Mortality", aggLevels="none", elasticity=1)
#> Checking scenarios...
#> Calculating impacts...
#> 		Formatting initial results...
#> 		Calculating temperature-driven scaled impacts...
#> Formatting results...
#> 
#> Finished.

### Set end year for analysis to 2110 -- messages user and returns a null value since default scenarios only have values out to 2100
run5 <- run_fredi(maxYear=2110)
#> Checking scenarios...
#> Calculating impacts...
#> 		Formatting initial results...
#> 		Calculating temperature-driven scaled impacts...
#> 		Calculating SLR-driven scaled impacts...
#> Formatting results...
#> Aggregating impacts...
#> 
#> Finished.

### Set end year for analysis to 2300 -- messages user and returns a null value since default scenarios only have values out to 2100)
run6 <- run_fredi(thru2300=TRUE)
#> Checking scenarios...
#> Calculating impacts...
#> 		Formatting initial results...
#> 		Calculating temperature-driven scaled impacts...
#> 		Calculating SLR-driven scaled impacts...
#> Formatting results...
#> Aggregating impacts...
#> 
#> Finished.