Calculate climate change impacts on socially vulnerable (SV) populations throughout the 21st century for available sectors

run_fredi_sv allows users to project annual average climate change impacts throughout the 21st century (2010-2090) for socially vulnerable (SV) populations for available sectors. Users can run run_fredi_sv() for individual sectors to generate annual physical impacts for SV populations. run_fredi_sv() can be run with default population and climate (temperature and sea level rise trajectories) or provide custom trajectories. The output of run_fredi_sv() is an R data frame object containing annual average physical impacts at five-year increments for the period 2010 to 2090. Users have the option to write outputs to Excel files that provide additional visualization of SV outputs.

run_fredi_sv(
  sector = NULL,
  driverInput = NULL,
  popInput = NULL,
  silent = TRUE,
  save = FALSE,
  outpath = getwd(),
  overwrite = FALSE,
  addDate = FALSE,
  .testing = FALSE
)

Arguments

sector

A character vector indicating for which sector(s) to run the FrEDI SV module (see get_sv_sectorInfo() for a list of available sectors).

driverInput

A data frame of up to four custom scenarios for drivers (temperature or global mean sea level rise). driverInput requires a data frame with columns of "year" and "scenario". The data frame must also include a third column: "temp_C" for temperature-driven sectors (containing temperature values in degrees Celsius of warming for the contiguous U.S.) or "slr_cm" for sea level rise (SLR)-driven sectors (containing values for global mean sea level rise in centimeters). Run get_sv_sectorInfo(gcmOnly=TRUE) to see temperature-driven sectors in the SV module and get_sv_sectorInfo(slrOnly=TRUE) to see SLR-driven scenarios. Users can also pass a data frame with all four columns ("year", "scenario", "temp_C", and "slr_cm"), in which case run_fredi_sv() determines whether to use the "temp_C" or "slr_cm" column as the driver trajectory based on the specified sector. Driver inputs for all scenarios should start in the year 2000 or earlier. All scenarios must include at least two non-missing values (especially values before or at 2000 and at or after 2090). If any required columns are missing, run_fredi_sv() will use the default temperature or sea level rise scenario from run_fredi(). If the data frame passed to driverInput has more than four unique scenarios, run_fredi_sv() will only run the first four scenarios.

popInput

A data frame containing regional population trajectories for each of the seven regions for the contiguous U.S. ("Midwest", "Northeast", "Northern Plains", "Northwest", "Southeast", "Southern Plains", "Southwest") as defined by the National Climate Assessment (NCA). The data frame passed to popInput should have columns "year", "region", and "reg_pop", which respectively contain values for year, NCA region name, and regional population. popInput only accepts a data frame with a single scenario; run_fredi_sv() uses the same population scenario for any and all driver scenarios in the data frame passed to driverInput. If popInput=NULL (default), run_fredi_sv() will use the default regional population trajectories.

silent

A logical (TRUE/FALSE) value indicating the level of messaging desired by the user (defaults to silent=TRUE).

save

A logical (TRUE/FALSE) value indicating whether to save the results to an Excel file (defaults to save=FALSE).

outpath

A character string indicating a file directory to save the Excel file. Defaults to the working directory, i.e. outpath=getwd(). If the directory specified by outpath does not exist, run_fredi_sv() will attempt to create the specified directory.

overwrite

A logical (TRUE/FALSE) value indicating whether to overwrite an existing Excel file if save=TRUE (defaults to overwrite=FALSE). If overwrite=FALSE, run_fredi_sv() will not automatically overwrite an existing Excel file; however, if a file exists and overwrite=FALSE, run_fredi_sv() will message the user and the user will have the option to overwrite the existing file. If overwrite=TRUE and the Excel file exists in the output directory, run_fredi_sv() will overwrite the existing file without messaging the user.

addDate

A logical (TRUE/FALSE) value indicating whether to add the date to the name of the output Excel file if save=TRUE (defaults to addDate=FALSE). If save=TRUE and addDate=TRUE, run_fredi_sv() will append the system date to the beginning of the name of the outputs Excel file using the format "%Y%m%d" (see base::format() and base::Sys.Date() for additional information).

Value

The output of run_fredi_sv() is an R data frame object containing NCA region-specific annual average physical impacts for socially vulnerable groups at five-year increments between 2010 and 2090. An optional output of run_fredi_sv() is an Excel file containing the output data frame with basic visualizations of sector outputs.

Details

run_fredi_sv() projects annual climate change impacts for socially vulnerable (SV) populations throughout the 21st century (2010-2090) for available sectors, using default or user-specified population, temperature, and sea level rise (SLR) trajectories. run_fredi_sv() is the main function for the FrEDI Social Vulnerability (SV) module in the FrEDI R package, described elsewhere (See https://epa.gov/cira/FrEDI for more information). The SV module extends the FrEDI framework to socially vulnerable populations using data underlying a 2021 U.S. Environmental Protection Agency (EPA) report on Climate Change and Social Vulnerability in the United States.

Users can run run_fredi_sv() to generate annual physical impacts for SV groups for individual sectors. When running run_fredi_sv(), users must specify one of the sectors in the SV module; use get_sv_sectorInfo() for a list of available sectors.

run_fredi_sv() can be run with default population and climate (temperature and SLR) trajectories or use run_fredi_sv() to run custom scenarios. Running run_fredi_sv() with custom climate scenarios requires passing a data frame of scenarios to the driverInput argument. run_fredi_sv() can also be run with a custom population scenario by passing a data frame of regional population trajectories to the popInput argument; unlike climate scenarios, run_fredi_sv() will only run a single scenario at a time.

driverInput can take a data frame containing up to four custom scenarios for drivers (temperature or global mean sea level rise). driverInput requires a data frame with columns of "year" and "scenario". The data frame must also include a third column: "temp_C" for temperature-driven sectors (containing temperature values in degrees Celsius of warming for the contiguous U.S.) or "slr_cm" for sea level rise (SLR)-driven sectors (containing values for global mean sea level rise in centimeters). Run get_sv_sectorInfo(gcmOnly=TRUE) to see temperature-driven sectors in the SV module and get_sv_sectorInfo(slrOnly=TRUE) to see SLR-driven scenarios. Users can also pass a data frame with all four columns ("year", "scenario", "temp_C", and "slr_cm"), in which case run_fredi_sv() determines whether to use the "temp_C" or "slr_cm" column as the driver trajectory based on the specified sector. If any required columns are missing, run_fredi_sv() will use the default temperature or sea level rise scenario from run_fredi(). If the data frame passed to driverInput has more than four unique scenarios, run_fredi_sv() will only run the first four scenarios.

Temperature inputs must be temperature change in degrees Celsius for the contiguous U.S. (use convertTemps() to convert global temperatures to CONUS temperatures before passing to driverInput) relative to a 1995 baseline (where 1995 is the central year of a 1986-2005 baseline period; values should start at zero in the year 1995).

Sea level rise inputs must be in centimeters relative to a 2000 baseline (i.e., values should start at zero in the year 2000). Driver inputs for all scenarios should start in the year 2000 or earlier. All scenarios must include at least two non-missing values (especially values before or at 2000 and at or after 2090).

popInput can take a data frame containing a single scenario with regional population trajectories for each of the seven regions for the contiguous U.S. ("Midwest", "Northeast", "Northern Plains", "Northwest", "Southeast", "Southern Plains", "Southwest") as defined by the National Climate Assessment (NCA). The data frame passed to popInput should have columns "year", "region", and "reg_pop", which respectively contain values for year, NCA region name, and regional population. popInput only accepts a data frame with a single scenario; run_fredi_sv() uses the same population scenario for any and all driver scenarios in the data frame passed to driverInput. If popInput=NULL (default), run_fredi_sv() will use the default regional population trajectories. The default regional population scenario is drawn from the Integrated Climate and Land Use Scenarios version 2 (ICLUSv2) model (Bierwagen et al, 2010; EPA 2017) under the Median variant projection of United Nations (United Nations, 2015). Note that the FrEDI SV default population scenario differs from the default population scenario used by run_fredi().

The output of run_fredi_sv() is an R data frame object containing NCA region-specific annual average physical impacts for socially vulnerable groups at five-year increments between 2010 and 2090. Users have the additional option to write results to an Excel file by setting save=TRUE; output Excel files provide basic visualizations of output data. Additional arguments provide more control over how the outputs are saved if save=TRUE:

outpath can be used to specify the directory in which to save an Excel output file. Defaults to outpath=getwd() (i.e., the working directory). If the directory specified by outpath does not exist, run_fredi_sv() will attempt to create the specified directory.

overwrite can be used to force run_fredi_sv() to overwrite an existing Excel file in the output directory. If overwrite=FALSE, run_fredi_sv() will not automatically overwrite an existing Excel file; however, if a file exists and overwrite=FALSE, run_fredi_sv() will message the user and the user will have the option to overwrite the existing file. If overwrite=TRUE and the Excel file exists in the output directory, run_fredi_sv() will overwrite the existing file without messaging the user.

addDate can be used to append the date to the output Excel file. If save=TRUE and addDate=TRUE, run_fredi_sv() will append the system date to the beginning of the name of the outputs Excel file using the format "%Y%m%d" (see base::format() and base::Sys.Date() for additional information).

References

Bierwagen, B., D. M. Theobald, C. R. Pyke, A. Choate, P. Groth, J. V. Thomas, and P. Morefield. 2010. “National housing and impervious surface scenarios for integrated climate impact assessments.” Proc. Natl. Acad. Sci. 107 (49): 20887–20892. https://doi.org/10.1073/pnas.1002096107.

EPA. 2017. Multi-Model Framework for Quantitative Sectoral Impacts Analysis: A technical report for the Fourth National Climate Assessment. U.S. Environmental Protection Agency, EPA 430-R-17-001.

EPA. 2021. Technical Documentation on the Framework for Evaluating Damages and Impacts (FrEDI). U.S. Environmental Protection Agency, EPA 430-R-21-004. Available at https://epa.gov/cira/FrEDI/.

EPA. 2021. Climate Change and Social Vulnerability in the United States: A Focus on Six Impacts. U.S. Environmental Protection Agency, EPA 430-R-21-003. Available at https://www.epa.gov/cira/social-vulnerability-report/.

United Nations. 2015. World population prospects: The 2015 revision. New York: United Nations, Department of Economic and Social Affairs, Population Division.

U.S. Global Change Research Program. 2015. Scenarios for the National Climate Assessment. Available at https://scenarios.globalchange.gov/regions_nca4.

Examples

### Run SV Module with defaults without specifying sector
df_sv <- run_fredi_sv()
#> [1] "Please select a sector: "
#> [1] "1. Air Quality - Childhood Asthma"
#> [1] "2. Air Quality - Premature Mortality"
#> [1] "3. Labor"
#> [1] "4. Extreme Temperature"
#> [1] "5. Roads"
#> [1] "6. High Tide Flooding and Traffic"
#> [1] "7. Coastal Properties"
#> Running FrEDI SV for sector 'Air Quality - Childhood Asthma':
#> Error in str_match(., col0): could not find function "str_match"

### Return a character vector with the names of all of the sectors in the FrEDI SV Module:
get_sv_sectorInfo()
#> [1] "Air Quality - Childhood Asthma"    "Air Quality - Premature Mortality"
#> [3] "Coastal Properties"                "Extreme Temperature"              
#> [5] "High Tide Flooding and Traffic"    "Labor"                            
#> [7] "Roads"                            

### Return a data frame of all of the sectors in the FrEDI SV Module (sector names and additional information)
get_sv_sectorInfo(description=T)
#>                              sector modelType      driverUnit
#> 1    Air Quality - Childhood Asthma       GCM degrees Celsius
#> 2 Air Quality - Premature Mortality       GCM degrees Celsius
#> 3                Coastal Properties       SLR              cm
#> 4               Extreme Temperature       GCM degrees Celsius
#> 5    High Tide Flooding and Traffic       SLR              cm
#> 6                             Labor       GCM degrees Celsius
#> 7                             Roads       GCM degrees Celsius
#>                              variants
#> 1                                 N/A
#> 2                                 N/A
#> 3 Without Adaptation, With Adaptation
#> 4                                 N/A
#> 5 Without Adaptation, With Adaptation
#> 6                                 N/A
#> 7 No Adaptation, Proactive Adaptation
#>                                        impactUnit
#> 1                          Childhood asthma cases
#> 2                             Premature Mortality
#> 3 Individuals threatened with total property loss
#> 4                                       Mortality
#> 5                                  Hours of delay
#> 6                             Hours of labor lost
#> 7                                  Hours of delay

### Run SV Module with defaults for "Coastal Properties" without saving
df_sv <- run_fredi_sv(sector="Coastal Properties")
#> Running FrEDI SV for sector 'Coastal Properties':
#> Error in str_match(., col0): could not find function "str_match"

### Run SV Module with defaults for "Extreme Temperature" without saving
df_sv <- run_fredi_sv(sector="Extreme Temperature")
#> Running FrEDI SV for sector 'Extreme Temperature':
#> Error in str_match(., col0): could not find function "str_match"

### Run SV Module with defaults for "Extreme Temperature" with saving and add date to file name
df_sv <- run_fredi_sv(sector="Extreme Temperature", save=T, addDate=T)
#> Running FrEDI SV for sector 'Extreme Temperature':
#> Error in str_match(., col0): could not find function "str_match"

### Load temperature scenarios
load(gcamScenarios)
#> Error in load(gcamScenarios): bad 'file' argument

### Load population scenario
load(popScenario)
#> Error in load(popScenario): bad 'file' argument

### Run SV Module for "Extreme Temperature" with custom population and temperature scenarios. Save and overwrite previous results
df_sv <- run_fredi_sv(sector="Extreme Temperature", driverInput = gcamScenarios, popInput = popScenario, save=T, addDate=T, overwrite = T)
#> Running FrEDI SV for sector 'Extreme Temperature':
#> Error in str_match(., col0): could not find function "str_match"